polkadot_runtime_parachains

Module assigner_coretime

The parachain coretime assignment module.

Handles scheduling of assignments coming from the coretime/broker chain. For on-demand assignments it relies on the separate on-demand assignment provider, where it forwards requests to.

CoreDescriptor contains pointers to the begin and the end of a list of schedules, together with the currently active assignments.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

PartsOf57600

Fraction expressed as a nominator with an assumed denominator of 57,600.