Type Alias polkadot_runtime_parachains::assigner_coretime::pallet::Module

pub type Module<T> = Pallet<T>;

๐Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn assign_core( core_idx: CoreIndex, begin: BlockNumberFor<T>, assignments: Vec<(CoreAssignment, PartsOf57600)>, end_hint: Option<BlockNumberFor<T>> ) -> Result<(), DispatchError>

Append another assignment for a core.

Important only appending is allowed. Meaning, all already existing assignments must have a begin smaller than the one passed here. This restriction exists, because it makes the insertion O(1) and the author could not think of a reason, why this restriction should be causing any problems. Inserting arbitrarily causes a DispatchError::DisallowedInsert error. This restriction could easily be lifted if need be and in fact an implementation is available here. The problem is that insertion complexity then depends on the size of the existing queue, which makes determining weights hard and could lead to issues like overweight blocks (at least in theory).