Function polkadot_runtime_parachains::assigner_on_demand::pallet::dispatchables::place_order_keep_alive

pub fn place_order_keep_alive<T: Config>(
    max_amount: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
    para_id: ParaId
)

Same as the place_order_allow_death call , but with a check that placing the order will not reap the account.

Parameters:

origin: The sender of the call, funds will be withdrawn from this account.

max_amount: The maximum balance to withdraw from the origin to place an order.

para_id: A ParaId the origin wants to provide blockspace for.

Errors:

InsufficientBalance: from the Currency implementation

InvalidParaId

QueueFull

SpotPriceHigherThanMaxAmount

NoOnDemandCores

Events:

SpotOrderPlaced

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::place_order_keep_alive.