Module polkadot_runtime_parachains::configuration

Configuration manager for the Polkadot runtime parachains logic.

Configuration can change only at session boundaries and is buffered until then.

Re-exports

pub use pallet::*;

Modules

migration

A module that is responsible for migration of storage.

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

HostConfiguration

All configuration of the runtime with respect to parachains and parathreads.

SessionChangeOutcome

A struct that holds the configuration that was active before the session change and optionally a configuration that became active after the session change.

TestWeightInfo

Enums

InconsistentError

Enumerates the possible inconsistencies of HostConfiguration.

Traits

WeightInfo