polkadot_runtime_parachains::configuration::migration::v13

Type Alias V12SchedulerParams

pub type V12SchedulerParams<BlockNumber> = SchedulerParams<BlockNumber>;

The v12 SchedulerParams, before the max_relay_parent_session_age field was added. This is identical to polkadot_primitives::v9::SchedulerParams.

Aliased Type

pub struct V12SchedulerParams<BlockNumber> {
    pub group_rotation_frequency: BlockNumber,
    pub paras_availability_period: BlockNumber,
    pub max_validators_per_core: Option<u32>,
    pub lookahead: u32,
    pub num_cores: u32,
    pub max_availability_timeouts: u32,
    pub on_demand_queue_max_size: u32,
    pub on_demand_target_queue_utilization: Perbill,
    pub on_demand_fee_variability: Perbill,
    pub on_demand_base_fee: u128,
    pub ttl: BlockNumber,
}

Fields
group_rotation_frequency: BlockNumber

How often parachain groups should be rotated across parachains.

Must be non-zero.

paras_availability_period: BlockNumber

Availability timeout for a block on a core, measured in blocks.

This is the maximum amount of blocks after a core became occupied that validators have time to make the block available.

This value only has effect on group rotations. If backers backed something at the end of their rotation, the occupied core affects the backing group that comes afterwards. We limit the effect one backing group can have on the next to paras_availability_period blocks.

Within a group rotation there is no timeout as backers are only affecting themselves.

Must be at least 1. With a value of 1, the previous group will not be able to negatively affect the following group at the expense of a tight availability timeline at group rotation boundaries.

max_validators_per_core: Option<u32>

The maximum number of validators to have per core.

None means no maximum.

lookahead: u32

The amount of blocks ahead to schedule paras.

num_cores: u32

How many cores are managed by the coretime chain.

max_availability_timeouts: u32
👎Deprecated

Deprecated and no longer used by the runtime. Removal is tracked by https://github.com/paritytech/polkadot-sdk/issues/6067.

on_demand_queue_max_size: u32

The maximum queue size of the pay as you go module.

on_demand_target_queue_utilization: Perbill

The target utilization of the spot price queue in percentages.

on_demand_fee_variability: Perbill

How quickly the fee rises in reaction to increased utilization. The lower the number the slower the increase.

on_demand_base_fee: u128

The minimum amount needed to claim a slot in the spot pricing queue.

ttl: BlockNumber
👎Deprecated

Deprecated and no longer used by the runtime. Removal is tracked by https://github.com/paritytech/polkadot-sdk/issues/6067.