Function polkadot_runtime_parachains::configuration::pallet::dispatchables::set_max_upward_message_num_per_candidate

pub fn set_max_upward_message_num_per_candidate<T: Config>(new: u32)

Sets the maximum number of messages that a candidate can contain.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_max_upward_message_num_per_candidate.