Enum polkadot_runtime_parachains::configuration::pallet::Call

pub enum Call<T: Config> {

    set_validation_upgrade_cooldown {
        new: T::BlockNumber,
    },
    set_validation_upgrade_delay {
        new: T::BlockNumber,
    },
    set_code_retention_period {
        new: T::BlockNumber,
    },
    set_max_code_size {
        new: u32,
    },
    set_max_pov_size {
        new: u32,
    },
    set_max_head_data_size {
        new: u32,
    },
    set_parathread_cores {
        new: u32,
    },
    set_parathread_retries {
        new: u32,
    },
    set_group_rotation_frequency {
        new: T::BlockNumber,
    },
    set_chain_availability_period {
        new: T::BlockNumber,
    },
    set_thread_availability_period {
        new: T::BlockNumber,
    },
    set_scheduling_lookahead {
        new: u32,
    },
    set_max_validators_per_core {
        new: Option<u32>,
    },
    set_max_validators {
        new: Option<u32>,
    },
    set_dispute_period {
        new: SessionIndex,
    },
    set_dispute_post_conclusion_acceptance_period {
        new: T::BlockNumber,
    },
    set_dispute_max_spam_slots {
        new: u32,
    },
    set_dispute_conclusion_by_time_out_period {
        new: T::BlockNumber,
    },
    set_no_show_slots {
        new: u32,
    },
    set_n_delay_tranches {
        new: u32,
    },
    set_zeroth_delay_tranche_width {
        new: u32,
    },
    set_needed_approvals {
        new: u32,
    },
    set_relay_vrf_modulo_samples {
        new: u32,
    },
    set_max_upward_queue_count {
        new: u32,
    },
    set_max_upward_queue_size {
        new: u32,
    },
    set_max_downward_message_size {
        new: u32,
    },
    set_ump_service_total_weight {
        new: Weight,
    },
    set_max_upward_message_size {
        new: u32,
    },
    set_max_upward_message_num_per_candidate {
        new: u32,
    },
    set_hrmp_open_request_ttl {
        new: u32,
    },
    set_hrmp_sender_deposit {
        new: Balance,
    },
    set_hrmp_recipient_deposit {
        new: Balance,
    },
    set_hrmp_channel_max_capacity {
        new: u32,
    },
    set_hrmp_channel_max_total_size {
        new: u32,
    },
    set_hrmp_max_parachain_inbound_channels {
        new: u32,
    },
    set_hrmp_max_parathread_inbound_channels {
        new: u32,
    },
    set_hrmp_channel_max_message_size {
        new: u32,
    },
    set_hrmp_max_parachain_outbound_channels {
        new: u32,
    },
    set_hrmp_max_parathread_outbound_channels {
        new: u32,
    },
    set_hrmp_max_message_num_per_candidate {
        new: u32,
    },
    set_ump_max_individual_weight {
        new: Weight,
    },
    set_pvf_checking_enabled {
        new: bool,
    },
    set_pvf_voting_ttl {
        new: SessionIndex,
    },
    set_minimum_validation_upgrade_delay {
        new: T::BlockNumber,
    },
    set_bypass_consistency_check {
        new: bool,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

set_validation_upgrade_cooldown

Fields

new: T::BlockNumber

Set the validation upgrade cooldown.

set_validation_upgrade_delay

Fields

new: T::BlockNumber

Set the validation upgrade delay.

set_code_retention_period

Fields

new: T::BlockNumber

Set the acceptance period for an included candidate.

set_max_code_size

Fields

new: u32

Set the max validation code size for incoming upgrades.

set_max_pov_size

Fields

new: u32

Set the max POV block size for incoming upgrades.

set_max_head_data_size

Fields

new: u32

Set the max head data size for paras.

set_parathread_cores

Fields

new: u32

Set the number of parathread execution cores.

set_parathread_retries

Fields

new: u32

Set the number of retries for a particular parathread.

set_group_rotation_frequency

Fields

new: T::BlockNumber

Set the parachain validator-group rotation frequency

set_chain_availability_period

Fields

new: T::BlockNumber

Set the availability period for parachains.

set_thread_availability_period

Fields

new: T::BlockNumber

Set the availability period for parathreads.

set_scheduling_lookahead

Fields

new: u32

Set the scheduling lookahead, in expected number of blocks at peak throughput.

set_max_validators_per_core

Fields

new: Option<u32>

Set the maximum number of validators to assign to any core.

set_max_validators

Fields

new: Option<u32>

Set the maximum number of validators to use in parachain consensus.

set_dispute_period

Fields

new: SessionIndex

Set the dispute period, in number of sessions to keep for disputes.

set_dispute_post_conclusion_acceptance_period

Fields

new: T::BlockNumber

Set the dispute post conclusion acceptance period.

set_dispute_max_spam_slots

Fields

new: u32

Set the maximum number of dispute spam slots.

set_dispute_conclusion_by_time_out_period

Fields

new: T::BlockNumber

Set the dispute conclusion by time out period.

set_no_show_slots

Fields

new: u32

Set the no show slots, in number of number of consensus slots. Must be at least 1.

set_n_delay_tranches

Fields

new: u32

Set the total number of delay tranches.

set_zeroth_delay_tranche_width

Fields

new: u32

Set the zeroth delay tranche width.

set_needed_approvals

Fields

new: u32

Set the number of validators needed to approve a block.

set_relay_vrf_modulo_samples

Fields

new: u32

Set the number of samples to do of the RelayVRFModulo approval assignment criterion.

set_max_upward_queue_count

Fields

new: u32

Sets the maximum items that can present in a upward dispatch queue at once.

set_max_upward_queue_size

Fields

new: u32

Sets the maximum total size of items that can present in a upward dispatch queue at once.

set_max_downward_message_size

Fields

new: u32

Set the critical downward message size.

set_ump_service_total_weight

Fields

new: Weight

Sets the soft limit for the phase of dispatching dispatchable upward messages.

set_max_upward_message_size

Fields

new: u32

Sets the maximum size of an upward message that can be sent by a candidate.

set_max_upward_message_num_per_candidate

Fields

new: u32

Sets the maximum number of messages that a candidate can contain.

set_hrmp_open_request_ttl

Fields

new: u32

Sets the number of sessions after which an HRMP open channel request expires.

set_hrmp_sender_deposit

Fields

new: Balance

Sets the amount of funds that the sender should provide for opening an HRMP channel.

set_hrmp_recipient_deposit

Fields

new: Balance

Sets the amount of funds that the recipient should provide for accepting opening an HRMP channel.

set_hrmp_channel_max_capacity

Fields

new: u32

Sets the maximum number of messages allowed in an HRMP channel at once.

set_hrmp_channel_max_total_size

Fields

new: u32

Sets the maximum total size of messages in bytes allowed in an HRMP channel at once.

set_hrmp_max_parachain_inbound_channels

Fields

new: u32

Sets the maximum number of inbound HRMP channels a parachain is allowed to accept.

set_hrmp_max_parathread_inbound_channels

Fields

new: u32

Sets the maximum number of inbound HRMP channels a parathread is allowed to accept.

set_hrmp_channel_max_message_size

Fields

new: u32

Sets the maximum size of a message that could ever be put into an HRMP channel.

set_hrmp_max_parachain_outbound_channels

Fields

new: u32

Sets the maximum number of outbound HRMP channels a parachain is allowed to open.

set_hrmp_max_parathread_outbound_channels

Fields

new: u32

Sets the maximum number of outbound HRMP channels a parathread is allowed to open.

set_hrmp_max_message_num_per_candidate

Fields

new: u32

Sets the maximum number of outbound HRMP messages can be sent by a candidate.

set_ump_max_individual_weight

Fields

new: Weight

Sets the maximum amount of weight any individual upward message may consume.

set_pvf_checking_enabled

Fields

new: bool

Enable or disable PVF pre-checking. Consult the field documentation prior executing.

set_pvf_voting_ttl

Fields

new: SessionIndex

Set the number of session changes after which a PVF pre-checking voting is rejected.

set_minimum_validation_upgrade_delay

Fields

new: T::BlockNumber

Sets the minimum delay between announcing the upgrade block for a parachain until the upgrade taking place.

See the field documentation for information and constraints for the new value.

set_bypass_consistency_check

Fields

new: bool

Setting this to true will disable consistency checks for the configuration setters. Use with caution.

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_set_validation_upgrade_cooldown(
new: T::BlockNumber
) -> Self

Create a call with the variant set_validation_upgrade_cooldown.

pub fn new_call_variant_set_validation_upgrade_delay(
new: T::BlockNumber
) -> Self

Create a call with the variant set_validation_upgrade_delay.

pub fn new_call_variant_set_code_retention_period(new: T::BlockNumber) -> Self

Create a call with the variant set_code_retention_period.

pub fn new_call_variant_set_max_code_size(new: u32) -> Self

Create a call with the variant set_max_code_size.

pub fn new_call_variant_set_max_pov_size(new: u32) -> Self

Create a call with the variant set_max_pov_size.

pub fn new_call_variant_set_max_head_data_size(new: u32) -> Self

Create a call with the variant set_max_head_data_size.

pub fn new_call_variant_set_parathread_cores(new: u32) -> Self

Create a call with the variant set_parathread_cores.

pub fn new_call_variant_set_parathread_retries(new: u32) -> Self

Create a call with the variant set_parathread_retries.

pub fn new_call_variant_set_group_rotation_frequency(
new: T::BlockNumber
) -> Self

Create a call with the variant set_group_rotation_frequency.

pub fn new_call_variant_set_chain_availability_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_chain_availability_period.

pub fn new_call_variant_set_thread_availability_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_thread_availability_period.

pub fn new_call_variant_set_scheduling_lookahead(new: u32) -> Self

Create a call with the variant set_scheduling_lookahead.

pub fn new_call_variant_set_max_validators_per_core(new: Option<u32>) -> Self

Create a call with the variant set_max_validators_per_core.

pub fn new_call_variant_set_max_validators(new: Option<u32>) -> Self

Create a call with the variant set_max_validators.

pub fn new_call_variant_set_dispute_period(new: SessionIndex) -> Self

Create a call with the variant set_dispute_period.

pub fn new_call_variant_set_dispute_post_conclusion_acceptance_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_dispute_post_conclusion_acceptance_period.

pub fn new_call_variant_set_dispute_max_spam_slots(new: u32) -> Self

Create a call with the variant set_dispute_max_spam_slots.

pub fn new_call_variant_set_dispute_conclusion_by_time_out_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_dispute_conclusion_by_time_out_period.

pub fn new_call_variant_set_no_show_slots(new: u32) -> Self

Create a call with the variant set_no_show_slots.

pub fn new_call_variant_set_n_delay_tranches(new: u32) -> Self

Create a call with the variant set_n_delay_tranches.

pub fn new_call_variant_set_zeroth_delay_tranche_width(new: u32) -> Self

Create a call with the variant set_zeroth_delay_tranche_width.

pub fn new_call_variant_set_needed_approvals(new: u32) -> Self

Create a call with the variant set_needed_approvals.

pub fn new_call_variant_set_relay_vrf_modulo_samples(new: u32) -> Self

Create a call with the variant set_relay_vrf_modulo_samples.

pub fn new_call_variant_set_max_upward_queue_count(new: u32) -> Self

Create a call with the variant set_max_upward_queue_count.

pub fn new_call_variant_set_max_upward_queue_size(new: u32) -> Self

Create a call with the variant set_max_upward_queue_size.

pub fn new_call_variant_set_max_downward_message_size(new: u32) -> Self

Create a call with the variant set_max_downward_message_size.

pub fn new_call_variant_set_ump_service_total_weight(new: Weight) -> Self

Create a call with the variant set_ump_service_total_weight.

pub fn new_call_variant_set_max_upward_message_size(new: u32) -> Self

Create a call with the variant set_max_upward_message_size.

pub fn new_call_variant_set_max_upward_message_num_per_candidate(
new: u32
) -> Self

Create a call with the variant set_max_upward_message_num_per_candidate.

pub fn new_call_variant_set_hrmp_open_request_ttl(new: u32) -> Self

Create a call with the variant set_hrmp_open_request_ttl.

pub fn new_call_variant_set_hrmp_sender_deposit(new: Balance) -> Self

Create a call with the variant set_hrmp_sender_deposit.

pub fn new_call_variant_set_hrmp_recipient_deposit(new: Balance) -> Self

Create a call with the variant set_hrmp_recipient_deposit.

pub fn new_call_variant_set_hrmp_channel_max_capacity(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_capacity.

pub fn new_call_variant_set_hrmp_channel_max_total_size(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_total_size.

pub fn new_call_variant_set_hrmp_max_parachain_inbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parachain_inbound_channels.

pub fn new_call_variant_set_hrmp_max_parathread_inbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parathread_inbound_channels.

pub fn new_call_variant_set_hrmp_channel_max_message_size(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_message_size.

pub fn new_call_variant_set_hrmp_max_parachain_outbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parachain_outbound_channels.

pub fn new_call_variant_set_hrmp_max_parathread_outbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parathread_outbound_channels.

pub fn new_call_variant_set_hrmp_max_message_num_per_candidate(new: u32) -> Self

Create a call with the variant set_hrmp_max_message_num_per_candidate.

pub fn new_call_variant_set_ump_max_individual_weight(new: Weight) -> Self

Create a call with the variant set_ump_max_individual_weight.

pub fn new_call_variant_set_pvf_checking_enabled(new: bool) -> Self

Create a call with the variant set_pvf_checking_enabled.

pub fn new_call_variant_set_pvf_voting_ttl(new: SessionIndex) -> Self

Create a call with the variant set_pvf_voting_ttl.

pub fn new_call_variant_set_minimum_validation_upgrade_delay(
new: T::BlockNumber
) -> Self

Create a call with the variant set_minimum_validation_upgrade_delay.

pub fn new_call_variant_set_bypass_consistency_check(new: bool) -> Self

Create a call with the variant set_bypass_consistency_check.