Struct polkadot_runtime_parachains::configuration::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn set_validation_upgrade_frequency(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the validation upgrade frequency.

pub fn set_validation_upgrade_delay(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the validation upgrade delay.

pub fn set_code_retention_period(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the acceptance period for an included candidate.

pub fn set_max_code_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max validation code size for incoming upgrades.

pub fn set_max_pov_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max POV block size for incoming upgrades.

pub fn set_max_head_data_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max head data size for paras.

pub fn set_parathread_cores(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of parathread execution cores.

pub fn set_parathread_retries(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of retries for a particular parathread.

pub fn set_group_rotation_frequency(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the parachain validator-group rotation frequency

pub fn set_chain_availability_period(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the availability period for parachains.

pub fn set_thread_availability_period(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the availability period for parathreads.

pub fn set_scheduling_lookahead(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Set the scheduling lookahead, in expected number of blocks at peak throughput.

pub fn set_max_validators_per_core(
origin: OriginFor<T>,
new: Option<u32>
) -> DispatchResult

Set the maximum number of validators to assign to any core.

pub fn set_max_validators(
origin: OriginFor<T>,
new: Option<u32>
) -> DispatchResult

Set the maximum number of validators to use in parachain consensus.

pub fn set_dispute_period(
origin: OriginFor<T>,
new: SessionIndex
) -> DispatchResult

Set the dispute period, in number of sessions to keep for disputes.

pub fn set_dispute_post_conclusion_acceptance_period(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the dispute post conclusion acceptance period.

pub fn set_dispute_max_spam_slots(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Set the maximum number of dispute spam slots.

pub fn set_dispute_conclusion_by_time_out_period(
origin: OriginFor<T>,
new: T::BlockNumber
) -> DispatchResult

Set the dispute conclusion by time out period.

pub fn set_no_show_slots(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the no show slots, in number of number of consensus slots. Must be at least 1.

pub fn set_n_delay_tranches(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the total number of delay tranches.

pub fn set_zeroth_delay_tranche_width(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Set the zeroth delay tranche width.

pub fn set_needed_approvals(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of validators needed to approve a block.

pub fn set_relay_vrf_modulo_samples(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Set the number of samples to do of the RelayVRFModulo approval assignment criterion.

pub fn set_max_upward_queue_count(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum items that can present in a upward dispatch queue at once.

pub fn set_max_upward_queue_size(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum total size of items that can present in a upward dispatch queue at once.

pub fn set_max_downward_message_size(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Set the critical downward message size.

pub fn set_ump_service_total_weight(
origin: OriginFor<T>,
new: Weight
) -> DispatchResult

Sets the soft limit for the phase of dispatching dispatchable upward messages.

pub fn set_max_upward_message_size(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum size of an upward message that can be sent by a candidate.

pub fn set_max_upward_message_num_per_candidate(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of messages that a candidate can contain.

pub fn set_hrmp_open_request_ttl(
_origin: OriginFor<T>,
_new: u32
) -> DispatchResult

Sets the number of sessions after which an HRMP open channel request expires.

pub fn set_hrmp_sender_deposit(
origin: OriginFor<T>,
new: Balance
) -> DispatchResult

Sets the amount of funds that the sender should provide for opening an HRMP channel.

pub fn set_hrmp_recipient_deposit(
origin: OriginFor<T>,
new: Balance
) -> DispatchResult

Sets the amount of funds that the recipient should provide for accepting opening an HRMP channel.

pub fn set_hrmp_channel_max_capacity(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of messages allowed in an HRMP channel at once.

pub fn set_hrmp_channel_max_total_size(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum total size of messages in bytes allowed in an HRMP channel at once.

pub fn set_hrmp_max_parachain_inbound_channels(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of inbound HRMP channels a parachain is allowed to accept.

pub fn set_hrmp_max_parathread_inbound_channels(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of inbound HRMP channels a parathread is allowed to accept.

pub fn set_hrmp_channel_max_message_size(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum size of a message that could ever be put into an HRMP channel.

pub fn set_hrmp_max_parachain_outbound_channels(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of outbound HRMP channels a parachain is allowed to open.

pub fn set_hrmp_max_parathread_outbound_channels(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of outbound HRMP channels a parathread is allowed to open.

pub fn set_hrmp_max_message_num_per_candidate(
origin: OriginFor<T>,
new: u32
) -> DispatchResult

Sets the maximum number of outbound HRMP messages can be sent by a candidate.

pub fn set_ump_max_individual_weight(
origin: OriginFor<T>,
new: Weight
) -> DispatchResult

Sets the maximum amount of weight any individual upward message may consume.

impl<T: Config> Pallet<T>

pub fn error_metadata() -> Option<PalletErrorMetadata>

impl<T: Config> Pallet<T>

pub fn config() -> HostConfiguration<T::BlockNumber>

The active configuration for the current session.

impl<T: Config> Pallet<T>

pub fn force_set_active_config(config: HostConfiguration<T::BlockNumber>)

Forcibly set the active config. This should be used with extreme care, and typically only when enabling parachains runtime modules for the first time on a chain which has been running without them.