Type Alias polkadot_runtime_parachains::configuration::pallet::Module

pub type Module<T> = Pallet<T>;

ðŸ‘ŽDeprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn set_validation_upgrade_cooldown( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the validation upgrade cooldown.

pub fn set_validation_upgrade_delay( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the validation upgrade delay.

pub fn set_code_retention_period( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the acceptance period for an included candidate.

pub fn set_max_code_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max validation code size for incoming upgrades.

pub fn set_max_pov_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max POV block size for incoming upgrades.

pub fn set_max_head_data_size(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the max head data size for paras.

pub fn set_coretime_cores(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of coretime execution cores.

Note that this configuration is managed by the coretime chain. Only manually change this, if you really know what you are doing!

pub fn set_on_demand_retries(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of retries for a particular on demand.

pub fn set_group_rotation_frequency( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the parachain validator-group rotation frequency

pub fn set_paras_availability_period( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the availability period for paras.

pub fn set_scheduling_lookahead( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the scheduling lookahead, in expected number of blocks at peak throughput.

pub fn set_max_validators_per_core( origin: OriginFor<T>, new: Option<u32> ) -> DispatchResult

Set the maximum number of validators to assign to any core.

pub fn set_max_validators( origin: OriginFor<T>, new: Option<u32> ) -> DispatchResult

Set the maximum number of validators to use in parachain consensus.

pub fn set_dispute_period( origin: OriginFor<T>, new: SessionIndex ) -> DispatchResult

Set the dispute period, in number of sessions to keep for disputes.

pub fn set_dispute_post_conclusion_acceptance_period( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the dispute post conclusion acceptance period.

pub fn set_no_show_slots(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the no show slots, in number of number of consensus slots. Must be at least 1.

pub fn set_n_delay_tranches(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the total number of delay tranches.

pub fn set_zeroth_delay_tranche_width( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the zeroth delay tranche width.

pub fn set_needed_approvals(origin: OriginFor<T>, new: u32) -> DispatchResult

Set the number of validators needed to approve a block.

pub fn set_relay_vrf_modulo_samples( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the number of samples to do of the RelayVRFModulo approval assignment criterion.

pub fn set_max_upward_queue_count( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum items that can present in a upward dispatch queue at once.

pub fn set_max_upward_queue_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum total size of items that can present in a upward dispatch queue at once.

pub fn set_max_downward_message_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the critical downward message size.

pub fn set_max_upward_message_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum size of an upward message that can be sent by a candidate.

pub fn set_max_upward_message_num_per_candidate( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum number of messages that a candidate can contain.

pub fn set_hrmp_open_request_ttl( _origin: OriginFor<T>, _new: u32 ) -> DispatchResult

Sets the number of sessions after which an HRMP open channel request expires.

pub fn set_hrmp_sender_deposit( origin: OriginFor<T>, new: Balance ) -> DispatchResult

Sets the amount of funds that the sender should provide for opening an HRMP channel.

pub fn set_hrmp_recipient_deposit( origin: OriginFor<T>, new: Balance ) -> DispatchResult

Sets the amount of funds that the recipient should provide for accepting opening an HRMP channel.

pub fn set_hrmp_channel_max_capacity( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum number of messages allowed in an HRMP channel at once.

pub fn set_hrmp_channel_max_total_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum total size of messages in bytes allowed in an HRMP channel at once.

pub fn set_hrmp_max_parachain_inbound_channels( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum number of inbound HRMP channels a parachain is allowed to accept.

pub fn set_hrmp_channel_max_message_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum size of a message that could ever be put into an HRMP channel.

pub fn set_hrmp_max_parachain_outbound_channels( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum number of outbound HRMP channels a parachain is allowed to open.

pub fn set_hrmp_max_message_num_per_candidate( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Sets the maximum number of outbound HRMP messages can be sent by a candidate.

pub fn set_pvf_voting_ttl( origin: OriginFor<T>, new: SessionIndex ) -> DispatchResult

Set the number of session changes after which a PVF pre-checking voting is rejected.

pub fn set_minimum_validation_upgrade_delay( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Sets the minimum delay between announcing the upgrade block for a parachain until the upgrade taking place.

See the field documentation for information and constraints for the new value.

pub fn set_bypass_consistency_check( origin: OriginFor<T>, new: bool ) -> DispatchResult

Setting this to true will disable consistency checks for the configuration setters. Use with caution.

pub fn set_async_backing_params( origin: OriginFor<T>, new: AsyncBackingParams ) -> DispatchResult

Set the asynchronous backing parameters.

pub fn set_executor_params( origin: OriginFor<T>, new: ExecutorParams ) -> DispatchResult

Set PVF executor parameters.

pub fn set_on_demand_base_fee( origin: OriginFor<T>, new: Balance ) -> DispatchResult

Set the on demand (parathreads) base fee.

pub fn set_on_demand_fee_variability( origin: OriginFor<T>, new: Perbill ) -> DispatchResult

Set the on demand (parathreads) fee variability.

pub fn set_on_demand_queue_max_size( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the on demand (parathreads) queue max size.

pub fn set_on_demand_target_queue_utilization( origin: OriginFor<T>, new: Perbill ) -> DispatchResult

Set the on demand (parathreads) fee variability.

pub fn set_on_demand_ttl( origin: OriginFor<T>, new: BlockNumberFor<T> ) -> DispatchResult

Set the on demand (parathreads) ttl in the claimqueue.

pub fn set_minimum_backing_votes( origin: OriginFor<T>, new: u32 ) -> DispatchResult

Set the minimum backing votes threshold.

pub fn set_node_feature( origin: OriginFor<T>, index: u8, value: bool ) -> DispatchResult

Set/Unset a node feature.

pub fn set_approval_voting_params( origin: OriginFor<T>, new: ApprovalVotingParams ) -> DispatchResult

Set approval-voting-params.

impl<T: Config> Pallet<T>

pub fn set_coretime_cores_unchecked(new: u32) -> DispatchResult

Set coretime cores.

To be used if authorization is checked otherwise.

impl<T: Config> Pallet<T>

pub fn config() -> HostConfiguration<BlockNumberFor<T>>

An auto-generated getter for ActiveConfig.

impl<T: Config> Pallet<T>

pub fn force_set_active_config(config: HostConfiguration<BlockNumberFor<T>>)

Forcibly set the active config. This should be used with extreme care, and typically only when enabling parachains runtime pallets for the first time on a chain which has been running without them.