Struct polkadot_runtime_parachains::configuration::HostConfiguration

pub struct HostConfiguration<BlockNumber> {

    pub max_code_size: u32,
    pub max_head_data_size: u32,
    pub max_upward_queue_count: u32,
    pub max_upward_queue_size: u32,
    pub max_upward_message_size: u32,
    pub max_upward_message_num_per_candidate: u32,
    pub hrmp_max_message_num_per_candidate: u32,
    pub validation_upgrade_cooldown: BlockNumber,
    pub validation_upgrade_delay: BlockNumber,
    pub async_backing_params: AsyncBackingParams,
    pub max_pov_size: u32,
    pub max_downward_message_size: u32,
    pub hrmp_max_parachain_outbound_channels: u32,
    pub hrmp_sender_deposit: Balance,
    pub hrmp_recipient_deposit: Balance,
    pub hrmp_channel_max_capacity: u32,
    pub hrmp_channel_max_total_size: u32,
    pub hrmp_max_parachain_inbound_channels: u32,
    pub hrmp_channel_max_message_size: u32,
    pub executor_params: ExecutorParams,
    pub code_retention_period: BlockNumber,
    pub on_demand_cores: u32,
    pub on_demand_retries: u32,
    pub on_demand_queue_max_size: u32,
    pub on_demand_target_queue_utilization: Perbill,
    pub on_demand_fee_variability: Perbill,
    pub on_demand_base_fee: Balance,
    pub on_demand_ttl: BlockNumber,
    pub group_rotation_frequency: BlockNumber,
    pub paras_availability_period: BlockNumber,
    pub scheduling_lookahead: u32,
    pub max_validators_per_core: Option<u32>,
    pub max_validators: Option<u32>,
    pub dispute_period: SessionIndex,
    pub dispute_post_conclusion_acceptance_period: BlockNumber,
    pub no_show_slots: u32,
    pub n_delay_tranches: u32,
    pub zeroth_delay_tranche_width: u32,
    pub needed_approvals: u32,
    pub relay_vrf_modulo_samples: u32,
    pub pvf_voting_ttl: SessionIndex,
    pub minimum_validation_upgrade_delay: BlockNumber,

}

All configuration of the runtime with respect to paras.

Fields
max_code_size: u32

The parameters that are required for the parachains. The maximum validation code size, in bytes.

max_head_data_size: u32

The maximum head-data size, in bytes.

max_upward_queue_count: u32

Total number of individual messages allowed in the parachain -> relay-chain message queue.

max_upward_queue_size: u32

Total size of messages allowed in the parachain -> relay-chain message queue before which no further messages may be added to it. If it exceeds this then the queue may contain only a single message.

max_upward_message_size: u32

The maximum size of an upward message that can be sent by a candidate.

This parameter affects the size upper bound of the CandidateCommitments.

max_upward_message_num_per_candidate: u32

The maximum number of messages that a candidate can contain.

This parameter affects the size upper bound of the CandidateCommitments.

hrmp_max_message_num_per_candidate: u32

The maximum number of outbound HRMP messages can be sent by a candidate.

This parameter affects the upper bound of size of CandidateCommitments.

validation_upgrade_cooldown: BlockNumber

The minimum period, in blocks, between which parachains can update their validation code.

This number is used to prevent parachains from spamming the relay chain with validation code upgrades. The only thing it controls is the number of blocks the UpgradeRestrictionSignal is set for the parachain in question.

If PVF pre-checking is enabled this should be greater than the maximum number of blocks PVF pre-checking can take. Intuitively, this number should be greater than the duration specified by [pvf_voting_ttl]. Unlike, [pvf_voting_ttl], this parameter uses blocks as a unit.

validation_upgrade_delay: BlockNumber

The delay, in blocks, after which an upgrade of the validation code is applied.

The upgrade for a parachain takes place when the first candidate which has relay-parent >= the relay-chain block where the upgrade is scheduled. This block is referred as to expected_at.

expected_at is determined when the upgrade is scheduled. This happens when the candidate that signals the upgrade is enacted. Right now, the relay-parent block number of the candidate scheduling the upgrade is used to determine the expected_at. This may change in the future with #4601.

When PVF pre-checking is enabled, the upgrade is scheduled only after the PVF pre-check has been completed.

Note, there are situations in which expected_at in the past. For example, if [paras_availability_period] is less than the delay set by this field or if PVF pre-check took more time than the delay. In such cases, the upgrade is further at the earliest possible time determined by [minimum_validation_upgrade_delay].

The rationale for this delay has to do with relay-chain reversions. In case there is an invalid candidate produced with the new version of the code, then the relay-chain can revert [validation_upgrade_delay] many blocks back and still find the new code in the storage by hash.

async_backing_params: AsyncBackingParams

Asynchronous backing parameters.

max_pov_size: u32

The parameters that are not essential, but still may be of interest for parachains. The maximum POV block size, in bytes.

max_downward_message_size: u32

The maximum size of a message that can be put in a downward message queue.

Since we require receiving at least one DMP message the obvious upper bound of the size is the PoV size. Of course, there is a lot of other different things that a parachain may decide to do with its PoV so this value in practice will be picked as a fraction of the PoV size.

hrmp_max_parachain_outbound_channels: u32

The maximum number of outbound HRMP channels a parachain is allowed to open.

hrmp_sender_deposit: Balance

The deposit that the sender should provide for opening an HRMP channel.

hrmp_recipient_deposit: Balance

The deposit that the recipient should provide for accepting opening an HRMP channel.

hrmp_channel_max_capacity: u32

The maximum number of messages allowed in an HRMP channel at once.

hrmp_channel_max_total_size: u32

The maximum total size of messages in bytes allowed in an HRMP channel at once.

hrmp_max_parachain_inbound_channels: u32

The maximum number of inbound HRMP channels a parachain is allowed to accept.

hrmp_channel_max_message_size: u32

The maximum size of a message that could ever be put into an HRMP channel.

This parameter affects the upper bound of size of CandidateCommitments.

executor_params: ExecutorParams

The executor environment parameters

code_retention_period: BlockNumber

Parameters that will unlikely be needed by parachains. How long to keep code on-chain, in blocks. This should be sufficiently long that disputes have concluded.

on_demand_cores: u32

The amount of execution cores to dedicate to on demand execution.

on_demand_retries: u32

The number of retries that a on demand author has to submit their block.

on_demand_queue_max_size: u32

The maximum queue size of the pay as you go module.

on_demand_target_queue_utilization: Perbill

The target utilization of the spot price queue in percentages.

on_demand_fee_variability: Perbill

How quickly the fee rises in reaction to increased utilization. The lower the number the slower the increase.

on_demand_base_fee: Balance

The minimum amount needed to claim a slot in the spot pricing queue.

on_demand_ttl: BlockNumber

The number of blocks an on demand claim stays in the scheduler’s claimqueue before getting cleared. This number should go reasonably higher than the number of blocks in the async backing lookahead.

group_rotation_frequency: BlockNumber

How often parachain groups should be rotated across parachains.

Must be non-zero.

paras_availability_period: BlockNumber

The availability period, in blocks. This is the amount of blocks after inclusion that validators have to make the block available and signal its availability to the chain.

Must be at least 1.

scheduling_lookahead: u32

The amount of blocks ahead to schedule paras.

max_validators_per_core: Option<u32>

The maximum number of validators to have per core.

None means no maximum.

max_validators: Option<u32>

The maximum number of validators to use for parachain consensus, period.

None means no maximum.

dispute_period: SessionIndex

The amount of sessions to keep for disputes.

dispute_post_conclusion_acceptance_period: BlockNumber

How long after dispute conclusion to accept statements.

no_show_slots: u32

The amount of consensus slots that must pass between submitting an assignment and submitting an approval vote before a validator is considered a no-show.

Must be at least 1.

n_delay_tranches: u32

The number of delay tranches in total.

zeroth_delay_tranche_width: u32

The width of the zeroth delay tranche for approval assignments. This many delay tranches beyond 0 are all consolidated to form a wide 0 tranche.

needed_approvals: u32

The number of validators needed to approve a block.

relay_vrf_modulo_samples: u32

The number of samples to do of the RelayVRFModulo approval assignment criterion.

pvf_voting_ttl: SessionIndex

If an active PVF pre-checking vote observes this many number of sessions it gets automatically rejected.

0 means PVF pre-checking will be rejected on the first observed session unless the voting gained supermajority before that the session change.

minimum_validation_upgrade_delay: BlockNumber

The lower bound number of blocks an upgrade can be scheduled.

Typically, upgrade gets scheduled [validation_upgrade_delay] relay-chain blocks after the relay-parent of the parablock that signalled the validation code upgrade. However, in the case a pre-checking voting was concluded in a longer duration the upgrade will be scheduled to the next block.

That can disrupt parachain inclusion. Specifically, it will make the blocks that were already backed invalid.

To prevent that, we introduce the minimum number of blocks after which the upgrade can be scheduled. This number is controlled by this field.

This value should be greater than [paras_availability_period].

Implementations

impl<BlockNumber> HostConfiguration<BlockNumber>where BlockNumber: Zero + PartialOrd + Debug + Clone + From<u32>,

pub fn check_consistency(&self) -> Result<(), InconsistentError<BlockNumber>>

Checks that this instance is consistent with the requirements on each individual member.

Errors

This function returns an error if the configuration is inconsistent.

pub fn panic_if_not_consistent(&self)

Checks that this instance is consistent with the requirements on each individual member.

Panics

This function panics if the configuration is inconsistent.