Trait polkadot_runtime_parachains::coretime::migration::GetLegacyLease

pub trait GetLegacyLease<N> {
    // Required methods
    fn get_parachain_lease_in_blocks(para: ParaId) -> Option<N>;
    fn get_all_parachains_with_leases() -> Vec<ParaId>;
}

Return information about a legacy lease of a parachain.

Required Methods

fn get_parachain_lease_in_blocks(para: ParaId) -> Option<N>

If parachain is a lease holding parachain, return the block at which the lease expires.

fn get_all_parachains_with_leases() -> Vec<ParaId>

Object Safety

This trait is not object safe.