polkadot_runtime_parachains::coretime::pallet::dispatchables

Function assign_core

pub fn assign_core<T: Config>(
    core: CoreIndex,
    begin: BlockNumberFor<T>,
    assignment: Vec<(CoreAssignment, PartsOf57600)>,
    end_hint: Option<BlockNumberFor<T>>,
)

Receive instructions from the ExternalBrokerOrigin, detailing how a specific core is to be used.

Parameters: -origin: The ExternalBrokerOrigin, assumed to be the coretime chain. -core: The core that should be scheduled. -begin: The starting blockheight of the instruction. -assignment: How the blockspace should be utilised. -end_hint: An optional hint as to when this particular set of instructions will end.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::assign_core.