Trait polkadot_runtime_parachains::coretime::pallet::Config

pub trait Config: Config + Config + Config {
    type RuntimeOrigin: From<<Self as Config>::RuntimeOrigin> + Into<Result<Origin, <Self as Config>::RuntimeOrigin>>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: Currency<Self::AccountId>;
    type BrokerId: Get<u32>;
    type BrokerPotLocation: Get<InteriorLocation>;
    type WeightInfo: WeightInfo;
    type SendXcm: SendXcm;
    type AssetTransactor: TransactAsset;
    type AccountToLocation: for<'a> TryConvert<&'a Self::AccountId, Location>;
    type MaxXcmTransactWeight: Get<Weight>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeOrigin: From<<Self as Config>::RuntimeOrigin> + Into<Result<Origin, <Self as Config>::RuntimeOrigin>>

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type Currency: Currency<Self::AccountId>

The runtime’s definition of a Currency.

type BrokerId: Get<u32>

The ParaId of the coretime chain.

type BrokerPotLocation: Get<InteriorLocation>

The coretime chain pot location.

type WeightInfo: WeightInfo

Something that provides the weight of this pallet.

type SendXcm: SendXcm

The XCM sender.

type AssetTransactor: TransactAsset

The asset transactor.

type AccountToLocation: for<'a> TryConvert<&'a Self::AccountId, Location>

AccountId to Location converter

type MaxXcmTransactWeight: Get<Weight>

Maximum weight for any XCM transact call that should be executed on the coretime chain.

Basically should be max_weight(set_leases, reserve, notify_core_count).

Object Safety

This trait is not object safe.