polkadot_runtime_parachains

Module disputes

Runtime component for handling disputes of parachain candidates.

Re-exports

pub use pallet::*;

Modules

migration

Storage migration(s) related to disputes pallet

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

slashing

Dispute slashing pallet.

Structs

TestWeightInfo

Enums

DisputeLocation

Whether the dispute is local or remote.

DisputeResult

The result of a dispute, whether the candidate is deemed valid (for) or invalid (against).

Traits

DisputesHandler

Hook into disputes handling.

RewardValidators

Reward hooks for disputes.

SlashingHandler

Punishment hooks for disputes.

WeightInfo