Type Alias polkadot_runtime_parachains::disputes::pallet::Frozen

pub type Frozen<T: Config> = StorageValue<_GeneratedPrefixForStorageFrozen<T>, Option<BlockNumberFor<T>>, ValueQuery>;

Whether the chain is frozen. Starts as None. When this is Some, the chain will not accept any new parachain blocks for backing or inclusion, and its value indicates the last valid block number in the chain. It can only be set back to None by governance intervention.

Storage type is StorageValue with value type Option < BlockNumberFor < T > >.

Aliased Type

struct Frozen<T: Config>(/* private fields */);