polkadot_runtime_parachains::hrmp::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

establish_channel_with_system

Establish a bidirectional HRMP channel between a parachain and a system chain.

establish_system_channel

Establish an HRMP channel between two system chains. If the channel does not already exist, the transaction fees will be refunded to the caller. The system does not take deposits for channels between system chains, and automatically sets the message number and size limits to the maximum allowed by the network’s configuration.

force_clean_hrmp

This extrinsic triggers the cleanup of all the HRMP storage items that a para may have. Normally this happens once per session, but this allows you to trigger the cleanup immediately for a specific parachain.

force_open_hrmp_channel

Open a channel from a sender to a recipient ParaId. Although opened by governance, the max_capacity and max_message_size are still subject to the Relay Chain’s configured limits.

force_process_hrmp_close

Force process HRMP close channel requests.

force_process_hrmp_open

Force process HRMP open channel requests.

hrmp_accept_open_channel

Accept a pending open channel request from the given sender.

hrmp_cancel_open_request

This cancels a pending open channel request. It can be canceled by either of the sender or the recipient for that request. The origin must be either of those.

hrmp_close_channel

Initiate unilateral closing of a channel. The origin must be either the sender or the recipient in the channel being closed.

hrmp_init_open_channel

Initiate opening a channel from a parachain to a given recipient with given channel parameters.

poke_channel_deposits

Update the deposits held for an HRMP channel to the latest Configuration. Channels with system chains do not require a deposit.