polkadot_runtime_parachains::hrmp::pallet::dispatchables

Function force_process_hrmp_open

pub fn force_process_hrmp_open<T: Config>(channels: u32)

Force process HRMP open channel requests.

If there are pending HRMP open channel requests, you can use this function to process all of those requests immediately.

Total number of opening channels must be provided as witness data.

Origin must be the ChannelManager.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::force_process_hrmp_open.