Function polkadot_runtime_parachains::hrmp::pallet::dispatchables::hrmp_cancel_open_request

pub fn hrmp_cancel_open_request<T: Config>(
    channel_id: HrmpChannelId,
    open_requests: u32
)

This cancels a pending open channel request. It can be canceled by either of the sender or the recipient for that request. The origin must be either of those.

The cancellation happens immediately. It is not possible to cancel the request if it is already accepted.

Total number of open requests (i.e. HrmpOpenChannelRequestsList) must be provided as witness data.

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the hrmp_cancel_open_request variant of the pallet Call enum. You should not attempt to call this function directly.