Trait polkadot_runtime_parachains::hrmp::pallet::Config

pub trait Config:
    Config
    + Config
    + Config
    + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type RuntimeOrigin: From<Origin> + From<<Self as Config>::RuntimeOrigin> + Into<Result<Origin, <Self as Config>::RuntimeOrigin>>;
    type ChannelManager: EnsureOrigin<<Self as Config>::RuntimeOrigin>;
    type Currency: ReservableCurrency<Self::AccountId>;
    type DefaultChannelSizeAndCapacityWithSystem: Get<(u32, u32)>;
    type VersionWrapper: WrapVersion;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The outer event type.

type RuntimeOrigin: From<Origin> + From<<Self as Config>::RuntimeOrigin> + Into<Result<Origin, <Self as Config>::RuntimeOrigin>>

type ChannelManager: EnsureOrigin<<Self as Config>::RuntimeOrigin>

The origin that can perform “force” actions on channels.

type Currency: ReservableCurrency<Self::AccountId>

An interface for reserving deposits for opening channels.

NOTE that this Currency instance will be charged with the amounts defined in the Configuration pallet. Specifically, that means that the Balance of the Currency implementation should be the same as Balance as used in the Configuration.

type DefaultChannelSizeAndCapacityWithSystem: Get<(u32, u32)>

The default channel size and capacity to use when opening a channel to a system parachain.

type VersionWrapper: WrapVersion

Means of converting an Xcm into a VersionedXcm. This pallet sends HRMP XCM notifications to the channel-related parachains, while the WrapVersion implementation attempts to wrap them into the most suitable XCM version for the destination parachain.

NOTE: For example, pallet_xcm provides an accurate implementation (recommended), or the default () implementation uses the latest XCM version for all parachains.

type WeightInfo: WeightInfo

Something that provides the weight of this pallet.

Object Safety

This trait is not object safe.