Type Definition polkadot_runtime_parachains::hrmp::pallet::HrmpIngressChannelsIndex

pub type HrmpIngressChannelsIndex<T: Config> = StorageMap<_GeneratedPrefixForStorageHrmpIngressChannelsIndex<T>, Twox64Concat, ParaId, Vec<ParaId>, ValueQuery>;

Ingress/egress indexes allow to find all the senders and receivers given the opposite side. I.e.

(a) ingress index allows to find all the senders for a given recipient. (b) egress index allows to find all the recipients for a given sender.

Invariants:

for each ingress index entry for P each item I in the index should present in HrmpChannels as (I, P).

for each egress index entry for P each item E in the index should present in HrmpChannels as (P, E).

there should be no other dangling channels in HrmpChannels.

the vectors are sorted.