Module polkadot_runtime_parachains::inclusion

The inclusion pallet is responsible for inclusion and availability of scheduled parachains.

It is responsible for carrying candidates from being backable to being backed, and then from backed to included.

Re-exports

pub use pallet::*;

Modules

migration

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

CandidatePendingAvailability

A backed candidate pending availability.

TestWeightInfo

Enums

AggregateMessageOrigin

Aggregate message origin for the MessageQueue pallet.

UmpQueueId

Identifies a UMP queue inside the MessageQueue pallet.

Constants

MAX_UPWARD_MESSAGE_SIZE_BOUND

Maximum value that config.max_upward_message_size can be set to.

Traits

QueueFootprinter

Reads the footprint of queues for a specific origin type.

RewardValidators

A hook for applying validator rewards

WeightInfo

Type Aliases

MaxUmpMessageLenOf

The maximal length of a UMP message.