Trait polkadot_runtime_parachains::inclusion::pallet::Config

pub trait Config:
    Config
    + Config
    + Config
    + Config
    + Config
    + Config
    + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type DisputesHandler: DisputesHandler<BlockNumberFor<Self>>;
    type RewardValidators: RewardValidators;
    type MessageQueue: EnqueueMessage<AggregateMessageOrigin>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

type DisputesHandler: DisputesHandler<BlockNumberFor<Self>>

type RewardValidators: RewardValidators

type MessageQueue: EnqueueMessage<AggregateMessageOrigin>

The system message queue.

The message queue provides general queueing and processing functionality. Currently it replaces the old UMP dispatch queue. Other use-cases can be implemented as well by adding new variants to AggregateMessageOrigin.

type WeightInfo: WeightInfo

Weight info for the calls of this pallet.

Object Safety

This trait is not object safe.