Trait polkadot_runtime_parachains::inclusion::WeightInfo

pub trait WeightInfo {
    // Required method
    fn enact_candidate(u: u32, h: u32, c: u32) -> Weight;
}

Required Methods

fn enact_candidate(u: u32, h: u32, c: u32) -> Weight

Weight for enact_candidate extrinsic given the number of sent messages (ump, hrmp) and whether there is a new code for a runtime upgrade.

NOTE: due to a shortcoming of the current benchmarking framework, we use u32 for the code upgrade, even though it is a bool.

Object Safety

This trait is not object safe.