Trait polkadot_runtime_parachains::initializer::pallet::Config

pub trait Config:
    Config
    + Config
    + Config
    + Config
    + Config
    + Config
    + Config
    + Config
    + Config
    + Config {
    type Randomness: Randomness<Self::Hash, BlockNumberFor<Self>>;
    type ForceOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>;
    type CoretimeOnNewSession: OnNewSession<BlockNumberFor<Self>>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type Randomness: Randomness<Self::Hash, BlockNumberFor<Self>>

A randomness beacon.

type ForceOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>

An origin which is allowed to force updates to parachains.

type CoretimeOnNewSession: OnNewSession<BlockNumberFor<Self>>

Temporary hack to call Coretime::on_new_session on chains that support Coretime or to disable it on the ones that don’t support it. Can be removed and replaced by a simple bound to coretime::Config once all chains support it.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Object Safety

This trait is not object safe.