Module polkadot_runtime_parachains::on_demand

The parachain on demand assignment module.

Implements a mechanism for taking in orders for on-demand parachain (previously parathreads) assignments. This module is not handled by the initializer but is instead instantiated in the construct_runtime macro.

The module currently limits parallel execution of blocks from the same ParaId via a core affinity mechanism. As long as there exists an affinity for a CoreIndex for a specific ParaId, orders for blockspace for that ParaId will only be assigned to that CoreIndex.

NOTE: Once we have elastic scaling implemented we might want to extend this module to support ignoring core affinity up to a certain extend. This should be opt-in though as the parachain needs to support multiple cores in the same block. If we want to enable a single parachain occupying multiple cores in on-demand, we will likely add a separate order type, where the intent can be made explicit.

Re-exports

pub use pallet::*;

Modules

migration

A module that is responsible for migration of storage.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

TestWeightInfo

A weight info that is only suitable for testing.

Traits

WeightInfo