Trait polkadot_runtime_parachains::on_demand::pallet::Config

pub trait Config:
    Config
    + Config
    + Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Currency: Currency<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type TrafficDefaultValue: Get<FixedU128>;
    type MaxHistoricalRevenue: Get<u32>;
    type PalletId: Get<PalletId>;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The runtime’s definition of an event.

type Currency: Currency<Self::AccountId>

The runtime’s definition of a Currency.

type WeightInfo: WeightInfo

Something that provides the weight of this pallet.

type TrafficDefaultValue: Get<FixedU128>

The default value for the spot traffic multiplier.

type MaxHistoricalRevenue: Get<u32>

The maximum number of blocks some historical revenue information stored for.

type PalletId: Get<PalletId>

Identifier for the internal revenue balance.

Object Safety

This trait is not object safe.