Module polkadot_runtime_parachains::paras::pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

The main components of this pallet are:

Pallet, which implements all of the dispatchable extrinsics of the pallet, among other public functions.

The subset of the functions that are dispatchable can be identified either in the dispatchables module or in the Call enum.

storage_types, which contains the list of all types that are representing a storage item. Otherwise, all storage items are listed among Type Definitions.

Config, which contains the configuration trait of this pallet.

Event and Error, which are listed among the Enums.

Re-exports

pub use __tt_error_token_97 as tt_error_token;

pub use __tt_default_parts_103 as tt_default_parts;

pub use __tt_extra_parts_103 as tt_extra_parts;

pub use __tt_default_parts_v2_103 as tt_default_parts_v2;

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Pallet

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Error

The Error enum of this pallet.

Event

The Event enum of this pallet

Traits

Config

Configuration trait of this pallet.

Type Aliases

ActionsQueue

The actions to perform during the start of a specific session index.

CodeByHash

Validation code stored by its hash.

CurrentCodeHash

The validation code hash of every live para.

FutureCodeHash

The actual future code hash of a para.

FutureCodeUpgrades

The block number at which the planned code change is expected for a parachain.

Heads

The head-data of every registered para.

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

MostRecentContext

The context (relay-chain block number) of the most recent parachain head.

Parachains

All lease holding parachains. Ordered ascending by ParaId. On demand parachains are not included.

PastCodeMeta

Past code of parachains. The parachains themselves may not be registered anymore, but we also keep their code on-chain for the same amount of time as outdated code to keep it available for approval checkers.

UpgradeRestrictionSignal

This is used by the relay-chain to communicate that there are restrictions for performing an upgrade for this parachain.