Module polkadot_runtime_parachains::paras::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

add_trusted_validation_code

Adds the validation code to the storage.

force_note_new_head

Note a new block head for para within the context of the current block.

force_queue_action

Put a parachain directly into the next session’s action queue. We can’t queue it any sooner than this without going into the initializer…

force_schedule_code_upgrade

Schedule an upgrade as if it was scheduled in the given relay parent block.

force_set_current_code

Set the storage for the parachain validation code immediately.

force_set_current_head

Set the storage for the current parachain head data immediately.

force_set_most_recent_context

Set the storage for the current parachain head data immediately.

include_pvf_check_statement

Includes a statement for a PVF pre-checking vote. Potentially, finalizes the vote and enacts the results if that was the last vote before achieving the supermajority.

poke_unused_validation_code

Remove the validation code from the storage iff the reference count is 0.