Function polkadot_runtime_parachains::paras::pallet::dispatchables::force_queue_action

pub fn force_queue_action<T: Config>(para: Id)

Put a parachain directly into the next session’s action queue. We can’t queue it any sooner than this without going into the initializer…

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::force_queue_action.