Module polkadot_runtime_parachains::paras::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

ActionsQueue

The actions to perform during the start of a specific session index.

CodeByHash

Validation code stored by its hash.

CodeByHashRefs

The number of reference on the validation code in CodeByHash storage.

CurrentCodeHash

The validation code hash of every live para.

FutureCodeHash

The actual future code hash of a para.

FutureCodeUpgrades

The block number at which the planned code change is expected for a para. The change will be applied after the first parablock for this ID included which executes in the context of a relay chain block with a number >= expected_at.

Heads

The head-data of every registered para.

MostRecentContext

The context (relay-chain block number) of the most recent parachain head.

ParaLifecycles

The current lifecycle of a all known Para IDs.

Parachains

All lease holding parachains. Ordered ascending by ParaId. On demand parachains are not included.

PastCodeHash

Actual past code hash, indicated by the para id as well as the block number at which it became outdated.

PastCodeMeta

Past code of parachains. The parachains themselves may not be registered anymore, but we also keep their code on-chain for the same amount of time as outdated code to keep it available for approval checkers.

PastCodePruning

Which paras have past code that needs pruning and the relay-chain block at which the code was replaced. Note that this is the actual height of the included block, not the expected height at which the code upgrade would be applied, although they may be equal. This is to ensure the entire acceptance period is covered, not an offset acceptance period starting from the time at which the parachain perceives a code upgrade as having occurred. Multiple entries for a single para are permitted. Ordered ascending by block number.

PvfActiveVoteList

The list of all currently active PVF votes. Auxiliary to PvfActiveVoteMap.

PvfActiveVoteMap

All currently active PVF pre-checking votes.

UpcomingParasGenesis

Upcoming paras instantiation arguments.

UpcomingUpgrades

The list of upcoming code upgrades. Each item is a pair of which para performs a code upgrade and at which relay-chain block it is expected at.

UpgradeCooldowns

The list of parachains that are awaiting for their upgrade restriction to cooldown.

UpgradeGoAheadSignal

This is used by the relay-chain to communicate to a parachain a go-ahead with in the upgrade procedure.

UpgradeRestrictionSignal

This is used by the relay-chain to communicate that there are restrictions for performing an upgrade for this parachain.