Type Alias polkadot_runtime_parachains::paras::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

pub fn force_set_current_code( origin: OriginFor<T>, para: ParaId, new_code: ValidationCode ) -> DispatchResult

Set the storage for the parachain validation code immediately.

pub fn force_set_current_head( origin: OriginFor<T>, para: ParaId, new_head: HeadData ) -> DispatchResult

Set the storage for the current parachain head data immediately.

pub fn force_schedule_code_upgrade( origin: OriginFor<T>, para: ParaId, new_code: ValidationCode, relay_parent_number: BlockNumberFor<T> ) -> DispatchResult

Schedule an upgrade as if it was scheduled in the given relay parent block.

pub fn force_note_new_head( origin: OriginFor<T>, para: ParaId, new_head: HeadData ) -> DispatchResult

Note a new block head for para within the context of the current block.

pub fn force_queue_action(origin: OriginFor<T>, para: ParaId) -> DispatchResult

Put a parachain directly into the next session’s action queue. We can’t queue it any sooner than this without going into the initializer…

pub fn add_trusted_validation_code( origin: OriginFor<T>, validation_code: ValidationCode ) -> DispatchResult

Adds the validation code to the storage.

The code will not be added if it is already present. Additionally, if PVF pre-checking is running for that code, it will be instantly accepted.

Otherwise, the code will be added into the storage. Note that the code will be added into storage with reference count 0. This is to account the fact that there are no users for this code yet. The caller will have to make sure that this code eventually gets used by some parachain or removed from the storage to avoid storage leaks. For the latter prefer to use the poke_unused_validation_code dispatchable to raw storage manipulation.

This function is mainly meant to be used for upgrading parachains that do not follow the go-ahead signal while the PVF pre-checking feature is enabled.

pub fn poke_unused_validation_code( origin: OriginFor<T>, validation_code_hash: ValidationCodeHash ) -> DispatchResult

Remove the validation code from the storage iff the reference count is 0.

This is better than removing the storage directly, because it will not remove the code that was suddenly got used by some parachain while this dispatchable was pending dispatching.

pub fn include_pvf_check_statement( origin: OriginFor<T>, stmt: PvfCheckStatement, signature: ValidatorSignature ) -> DispatchResultWithPostInfo

Includes a statement for a PVF pre-checking vote. Potentially, finalizes the vote and enacts the results if that was the last vote before achieving the supermajority.

pub fn force_set_most_recent_context( origin: OriginFor<T>, para: ParaId, context: BlockNumberFor<T> ) -> DispatchResult

Set the storage for the current parachain head data immediately.

impl<T: Config> Pallet<T>

pub fn parachains() -> Vec<ParaId>

An auto-generated getter for Parachains.

impl<T: Config> Pallet<T>

pub fn para_head<KArg>(k: KArg) -> Option<HeadData>where KArg: EncodeLike<ParaId>,

An auto-generated getter for Heads.

impl<T: Config> Pallet<T>

pub fn para_most_recent_context<KArg>(k: KArg) -> Option<BlockNumberFor<T>>where KArg: EncodeLike<ParaId>,

An auto-generated getter for MostRecentContext.

impl<T: Config> Pallet<T>

pub fn current_code_hash<KArg>(k: KArg) -> Option<ValidationCodeHash>where KArg: EncodeLike<ParaId>,

An auto-generated getter for CurrentCodeHash.

impl<T: Config> Pallet<T>

pub fn past_code_meta<KArg>(k: KArg) -> ParaPastCodeMeta<BlockNumberFor<T>>where KArg: EncodeLike<ParaId>,

An auto-generated getter for PastCodeMeta.

impl<T: Config> Pallet<T>

pub fn future_code_upgrade_at<KArg>(k: KArg) -> Option<BlockNumberFor<T>>where KArg: EncodeLike<ParaId>,

An auto-generated getter for FutureCodeUpgrades.

impl<T: Config> Pallet<T>

pub fn future_code_hash<KArg>(k: KArg) -> Option<ValidationCodeHash>where KArg: EncodeLike<ParaId>,

An auto-generated getter for FutureCodeHash.

impl<T: Config> Pallet<T>

pub fn upgrade_restriction_signal<KArg>(k: KArg) -> Option<UpgradeRestriction>where KArg: EncodeLike<ParaId>,

An auto-generated getter for UpgradeRestrictionSignal.

impl<T: Config> Pallet<T>

pub fn actions_queue<KArg>(k: KArg) -> Vec<ParaId>where KArg: EncodeLike<SessionIndex>,

An auto-generated getter for ActionsQueue.

impl<T: Config> Pallet<T>

pub fn code_by_hash<KArg>(k: KArg) -> Option<ValidationCode>where KArg: EncodeLike<ValidationCodeHash>,

An auto-generated getter for CodeByHash.

impl<T: Config> Pallet<T>

pub fn can_schedule_para_initialize(id: &ParaId) -> bool

Verify that schedule_para_initialize can be called successfully.

Returns false if para is already registered in the system.

pub fn lifecycle(id: ParaId) -> Option<ParaLifecycle>

Returns the current lifecycle state of the para.

pub fn is_valid_para(id: ParaId) -> bool

Returns whether the given ID refers to a valid para.

Paras that are onboarding or offboarding are not included.

pub fn is_offboarding(id: ParaId) -> bool

Returns whether the given ID refers to a para that is offboarding.

An invalid or non-offboarding para ID will return false.

pub fn is_parachain(id: ParaId) -> bool

Whether a para ID corresponds to any live lease holding parachain.

Includes lease holding parachains which will downgrade to a on-demand parachains in the future.

pub fn is_parathread(id: ParaId) -> bool

Whether a para ID corresponds to any live parathread (on-demand parachain).

Includes on-demand parachains which will upgrade to lease holding parachains in the future.

pub fn test_on_new_session()

Test function for triggering a new session in this pallet.