polkadot_runtime_parachains::paras

Trait AssignCoretime

pub trait AssignCoretime {
    // Required method
    fn assign_coretime(id: ParaId) -> DispatchResult;
}

Assign coretime to some parachain.

This assigns coretime to a parachain without using the coretime chain. Thus, this should only be used for testing purposes.

Required Methods

fn assign_coretime(id: ParaId) -> DispatchResult

ONLY USE FOR TESTING OR GENESIS.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.