polkadot_runtime_parachains

Module paras_inherent

Provides glue code over the scheduler and inclusion modules, and accepting one inherent per block that can include new para candidates and bitfields.

Unlike other modules in this crate, it does not need to be initialized by the initializer, as it has no initialization logic and its finalization logic depends only on the details of this module.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

TestWeightInfo

Traits

IndexedRetain

A helper trait to allow calling retain while getting access to the index of the item in the vec.

IsSortedBy

Helper trait until is_sorted_by is stabilized. TODO: https://github.com/rust-lang/rust/issues/53485

WeightInfo

Functions

backed_candidate_weight

backed_candidates_weight

dispute_statement_set_weight

Get time weights from benchmarks and set proof size to tx size.

multi_dispute_statement_sets_weight

paras_inherent_total_weight

signed_bitfield_weight

signed_bitfields_weight