Trait polkadot_runtime_parachains::paras_inherent::pallet::Config

pub trait Config:
    Config
    + Config
    + Config
    + Config {
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Object Safety

This trait is not object safe.