Module polkadot_runtime_parachains::reward_points

An implementation of the RewardValidators trait used by inclusion that employs pallet-staking to compute the rewards.

Based on https://research.web3.foundation/en/latest/polkadot/overview/2-token-economics.html which doesn’t currently mention availability bitfields. As such, we don’t reward them for the time being, although we will build schemes to do so in the future.

Structs

RewardValidatorsWithEraPoints

Rewards validators for participating in parachains with era points in pallet-staking.

Constants

BACKING_POINTS

The amount of era points given by backing a candidate that is included.

DISPUTE_STATEMENT_POINTS

The amount of era points given by dispute voting on a candidate.