polkadot_runtime_parachains

Module runtime_api_impl

Runtime API implementations for Parachains.

These are exposed as different modules using different sets of primitives. At the moment there is a v2 module for the current stable API and vstaging module for all staging methods. When new version of the stable API is released it will be based on v2 and will contain methods from vstaging. The promotion consists of the following steps:

Bump the version of the stable module (e.g. v2 becomes v3)

Move methods from vstaging to v3. The new stable version should include all methods from vstaging tagged with the new version number (e.g. all v3 methods).

Modules

v11

A module exporting runtime API implementation functions for all runtime APIs using v5 primitives.

vstaging

Put implementations of functions from staging APIs here.