Module polkadot_runtime_parachains::runtime_api_impl::v10

A module exporting runtime API implementation functions for all runtime APIs using v5 primitives.

Runtimes implementing the v10 runtime API are recommended to forward directly to these functions.

Functions

approval_voting_params

Approval voting subsystem configuration parameters

assumed_validation_data

Implementation for the assumed_validation_data function of the runtime API.

async_backing_params

Implementation for AsyncBackingParams function from the runtime API

availability_cores

Implementation for the availability_cores function of the runtime API.

backing_state

Implementation for ParaBackingState function from the runtime API

candidate_events

Implementation for the candidate_events function of the runtime API.

candidate_pending_availability

Implementation for the candidate_pending_availability function of the runtime API.

check_validation_outputs

Implementation for the check_validation_outputs function of the runtime API.

disabled_validators

Implementation for DisabledValidators

dmq_contents

Implementation for the dmq_contents function of the runtime API.

get_session_disputes

Implementation for get_session_disputes function from the runtime API

inbound_hrmp_channels_contents

Implementation for the inbound_hrmp_channels_contents function of the runtime API.

minimum_backing_votes

Return the min backing votes threshold from the configuration.

node_features

Returns the current state of the node features.

on_chain_votes

Disputes imported via means of on-chain imports.

persisted_validation_data

Implementation for the persisted_validation_data function of the runtime API.

pvfs_require_precheck

Returns the list of all PVF code hashes that require pre-checking.

relevant_authority_ids

Implementation for the AuthorityDiscoveryApi::authorities() function of the runtime API. It is a heavy call, but currently only used for authority discovery, so it is fine. Gets next, current and some historical authority ids using session_info module.

session_executor_params

Get session executor parameter set

session_index_for_child

Implementation for the session_index_for_child function of the runtime API.

session_info

Get the session info for the given session, if stored.

submit_pvf_check_statement

Submits an PVF pre-checking vote.

submit_unsigned_slashing_report

Implementation of submit_report_dispute_lost runtime API

unapplied_slashes

Implementation of unapplied_slashes runtime API

validation_code

Implementation for the validation_code function of the runtime API.

validation_code_by_hash

Implementation for the validation_code_by_hash function of the runtime API.

validation_code_hash

Returns the validation code hash for the given parachain making the given OccupiedCoreAssumption.

validator_groups

Implementation for the validator_groups function of the runtime API.

validators

Implementation for the validators function of the runtime API.