polkadot_runtime_parachains::scheduler::common

Trait AssignmentProvider

pub trait AssignmentProvider<BlockNumber> {
    // Required methods
    fn pop_assignment_for_core(core_idx: CoreIndex) -> Option<Assignment>;
    fn report_processed(assignment: Assignment);
    fn push_back_assignment(assignment: Assignment);
    fn assignment_duplicated(assignment: &Assignment);
}

Required Methods

fn pop_assignment_for_core(core_idx: CoreIndex) -> Option<Assignment>

Pops an Assignment from the provider for a specified CoreIndex.

This is where assignments come into existence.

fn report_processed(assignment: Assignment)

A previously popped Assignment has been fully processed.

Report back to the assignment provider that an assignment is done and no longer present in the scheduler.

This is one way of the life of an assignment coming to an end.

fn push_back_assignment(assignment: Assignment)

Push back a previously popped assignment.

If the assignment could not be processed within the current session, it can be pushed back to the assignment provider in order to be popped again later.

This is the second way the life of an assignment can come to an end.

fn assignment_duplicated(assignment: &Assignment)

Report that an assignment was duplicated by the scheduler.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.