Type Alias polkadot_runtime_parachains::scheduler::pallet::ParasEntryType

pub type ParasEntryType<T> = ParasEntry<BlockNumberFor<T>>;

Convenience type declaration for ParasEntry.

Aliased Type

struct ParasEntryType<T> {
    pub assignment: Assignment,
    pub availability_timeouts: u32,
    pub ttl: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
}

Fields
assignment: Assignment

The underlying Assignment.

availability_timeouts: u32

The number of times the entry has timed out in availability already.

ttl: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

The block height until this entry needs to be backed.

If missed the entry will be removed from the claim queue without ever having occupied the core.