polkadot_runtime_parachains

Module session_info

The session info pallet provides information about validator sets from prior sessions needed for approvals and disputes.

See the documentation on session info in the implementers’ guide.

Re-exports

pub use pallet::*;

Modules

migration

A module that is responsible for migration of storage.

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Traits

AuthorityDiscoveryConfig

An abstraction for the authority discovery pallet to help with mock testing.

Type Aliases

AccountId

A type for representing the validator account id in a session.

IdentificationTuple

A tuple of (AccountId, Identification) where Identification is the full identification of AccountId.