polkadot_runtime_parachains::session_info::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

AccountKeys

The validator account keys of the validators actively participating in parachain consensus.

AssignmentKeysUnsafe

Assignment keys for the current session. Note that this API is private due to it being prone to ‘off-by-one’ at session boundaries. When in doubt, use Sessions API instead.

EarliestStoredSession

The earliest session for which previous session info is stored.

SessionExecutorParams

Executor parameter set for a given session index

Sessions

Session information in a rolling window. Should have an entry in range EarliestStoredSession..=CurrentSessionIndex. Does not have any entries before the session index in the first session change notification.