polkadot_runtime_parachains::session_info

Trait AuthorityDiscoveryConfig

pub trait AuthorityDiscoveryConfig {
    // Required method
    fn authorities() -> Vec<AuthorityDiscoveryId> ;
}

An abstraction for the authority discovery pallet to help with mock testing.

Required Methods

fn authorities() -> Vec<AuthorityDiscoveryId>

Retrieve authority identifiers of the current authority set in canonical ordering.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.