Module polkadot_runtime_parachains::ump

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

XcmSink
A specific implementation of a UmpSink where messages are in the XCM format and will be forwarded to the XCM Executor.

Enums

AcceptanceCheckErr
An error returned by [check_upward_messages] that indicates a violation of one of acceptance criteria rules.

Traits

UmpSink
All upward messages coming from parachains will be funneled into an implementation of this trait.

Type Definitions

MessageId
Simple type used to identify messages for the purpose of reporting events. Secure if and only if the message content is unique.