Crate polkadot_runtime_parachains

Runtime modules for parachains code.

It is crucial to include all the modules from this crate in the runtime, in particular the Initializer module, as it is responsible for initializing the state of the other modules.

Re-exports

pub use origin::ensure_parachain;

pub use origin::Origin;

pub use paras::ParaLifecycle;

pub use paras::UpgradeStrategy;

Modules

assigner_coretime

The parachain coretime assignment module.

assigner_parachains

The bulk (parachain slot auction) blockspace assignment provider. This provider is tightly coupled with the configuration and paras modules.

configuration

Configuration manager for the Polkadot runtime parachains logic.

coretime

Extrinsics implementing the relay chain side of the Coretime interface.

disputes

Runtime component for handling disputes of parachain candidates.

dmp

To prevent Out of Memory errors on the DownwardMessageQueue, an exponential fee factor (DeliveryFeeFactor) is set. The fee factor increments exponentially after the number of messages in the DownwardMessageQueue passes a threshold. This threshold is set as:

hrmp

inclusion

The inclusion pallet is responsible for inclusion and availability of scheduled parachains.

initializer

This module is responsible for maintaining a consistent initialization order for all other parachains modules. It’s also responsible for finalization and session change notifications.

metrics

Runtime declaration of the parachain metrics.

on_demand

The parachain on demand assignment module.

origin

Declaration of the parachain specific origin and a pallet that hosts it.

paras

The paras pallet acts as the main registry of paras.

paras_inherent

Provides glue code over the scheduler and inclusion modules, and accepting one inherent per block that can include new para candidates and bitfields.

reward_points

An implementation of the RewardValidators trait used by inclusion that employs pallet-staking to compute the rewards.

runtime_api_impl

Runtime API implementations for Parachains.

scheduler

The scheduler module for parachains and parathreads.

session_info

The session info pallet provides information about validator sets from prior sessions needed for approvals and disputes.

shared

A pallet for any shared state that other pallets may want access to.

Traits

FeeTracker

Trait for tracking message delivery fees on a transport protocol.

Functions

schedule_code_upgrade

Schedules a validation code upgrade to a parachain with the given id.

schedule_para_cleanup

Schedule a para to be cleaned up at the start of the next session.

schedule_para_initialize

Schedule a para to be initialized at the start of the next session with the given genesis data.

schedule_parachain_downgrade

Schedule a lease holding parachain to be downgraded to an on-demand parachain.

schedule_parathread_upgrade

Schedule a parathread (on-demand parachain) to be upgraded to a lease holding parachain.

set_current_head

Sets the current parachain head with the given id.