Crate polkadot_runtime_parachains

Runtime modules for parachains code.

It is crucial to include all the modules from this crate in the runtime, in particular the Initializer module, as it is responsible for initializing the state of the other modules.

Re-exports

pub use origin::ensure_parachain;

pub use origin::Origin;

pub use paras::ParaLifecycle;

Modules

configuration

Configuration manager for the Polkadot runtime parachains logic.

disputes

Runtime component for handling disputes of parachain candidates.

dmp

hrmp

inclusion

The inclusion pallet is responsible for inclusion and availability of scheduled parachains and parathreads.

initializer

This module is responsible for maintaining a consistent initialization order for all other parachains modules. It’s also responsible for finalization and session change notifications.

origin

Declaration of the parachain specific origin and a pallet that hosts it.

paras

The paras pallet is responsible for storing data on parachains and parathreads.

paras_inherent

Provides glue code over the scheduler and inclusion modules, and accepting one inherent per block that can include new para candidates and bitfields.

reward_points

An implementation of the RewardValidators trait used by inclusion that employs pallet-staking to compute the rewards.

runtime_api_impl

Runtime API implementations for Parachains.

scheduler

The scheduler module for parachains and parathreads.

session_info

The session info pallet provides information about validator sets from prior sessions needed for approvals and disputes.

shared

A pallet for any shared state that other pallets may want access to.

ump

Functions

schedule_para_cleanup

Schedule a para to be cleaned up at the start of the next session.

schedule_para_initialize

Schedule a para to be initialized at the start of the next session with the given genesis data.

schedule_parachain_downgrade

Schedule a parachain to be downgraded to a parathread.

schedule_parathread_upgrade

Schedule a parathread to be upgraded to a parachain.