Module polkadot_runtime_parachains::configuration::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

set_bypass_consistency_check

Setting this to true will disable consistency checks for the configuration setters. Use with caution.

set_chain_availability_period

Set the availability period for parachains.

set_code_retention_period

Set the acceptance period for an included candidate.

set_dispute_conclusion_by_time_out_period

Set the dispute conclusion by time out period.

set_dispute_period

Set the dispute period, in number of sessions to keep for disputes.

set_dispute_post_conclusion_acceptance_period

Set the dispute post conclusion acceptance period.

set_group_rotation_frequency

Set the parachain validator-group rotation frequency

set_hrmp_channel_max_capacity

Sets the maximum number of messages allowed in an HRMP channel at once.

set_hrmp_channel_max_message_size

Sets the maximum size of a message that could ever be put into an HRMP channel.

set_hrmp_channel_max_total_size

Sets the maximum total size of messages in bytes allowed in an HRMP channel at once.

set_hrmp_max_message_num_per_candidate

Sets the maximum number of outbound HRMP messages can be sent by a candidate.

set_hrmp_max_parachain_inbound_channels

Sets the maximum number of inbound HRMP channels a parachain is allowed to accept.

set_hrmp_max_parachain_outbound_channels

Sets the maximum number of outbound HRMP channels a parachain is allowed to open.

set_hrmp_max_parathread_inbound_channels

Sets the maximum number of inbound HRMP channels a parathread is allowed to accept.

set_hrmp_max_parathread_outbound_channels

Sets the maximum number of outbound HRMP channels a parathread is allowed to open.

set_hrmp_open_request_ttl

Sets the number of sessions after which an HRMP open channel request expires.

set_hrmp_recipient_deposit

Sets the amount of funds that the recipient should provide for accepting opening an HRMP channel.

set_hrmp_sender_deposit

Sets the amount of funds that the sender should provide for opening an HRMP channel.

set_max_code_size

Set the max validation code size for incoming upgrades.

set_max_downward_message_size

Set the critical downward message size.

set_max_head_data_size

Set the max head data size for paras.

set_max_pov_size

Set the max POV block size for incoming upgrades.

set_max_upward_message_num_per_candidate

Sets the maximum number of messages that a candidate can contain.

set_max_upward_message_size

Sets the maximum size of an upward message that can be sent by a candidate.

set_max_upward_queue_count

Sets the maximum items that can present in a upward dispatch queue at once.

set_max_upward_queue_size

Sets the maximum total size of items that can present in a upward dispatch queue at once.

set_max_validators

Set the maximum number of validators to use in parachain consensus.

set_max_validators_per_core

Set the maximum number of validators to assign to any core.

set_minimum_validation_upgrade_delay

Sets the minimum delay between announcing the upgrade block for a parachain until the upgrade taking place.

set_n_delay_tranches

Set the total number of delay tranches.

set_needed_approvals

Set the number of validators needed to approve a block.

set_no_show_slots

Set the no show slots, in number of number of consensus slots. Must be at least 1.

set_parathread_cores

Set the number of parathread execution cores.

set_parathread_retries

Set the number of retries for a particular parathread.

set_pvf_checking_enabled

Enable or disable PVF pre-checking. Consult the field documentation prior executing.

set_pvf_voting_ttl

Set the number of session changes after which a PVF pre-checking voting is rejected.

set_relay_vrf_modulo_samples

Set the number of samples to do of the RelayVRFModulo approval assignment criterion.

set_scheduling_lookahead

Set the scheduling lookahead, in expected number of blocks at peak throughput.

set_thread_availability_period

Set the availability period for parathreads.

set_ump_max_individual_weight

Sets the maximum amount of weight any individual upward message may consume.

set_ump_service_total_weight

Sets the soft limit for the phase of dispatching dispatchable upward messages.

set_validation_upgrade_cooldown

Set the validation upgrade cooldown.

set_validation_upgrade_delay

Set the validation upgrade delay.

set_zeroth_delay_tranche_width

Set the zeroth delay tranche width.