Enum polkadot_runtime_parachains::configuration::pallet::Call

pub enum Call<T: Config> {

    set_validation_upgrade_frequency {
        new: T::BlockNumber,
    },
    set_validation_upgrade_delay {
        new: T::BlockNumber,
    },
    set_code_retention_period {
        new: T::BlockNumber,
    },
    set_max_code_size {
        new: u32,
    },
    set_max_pov_size {
        new: u32,
    },
    set_max_head_data_size {
        new: u32,
    },
    set_parathread_cores {
        new: u32,
    },
    set_parathread_retries {
        new: u32,
    },
    set_group_rotation_frequency {
        new: T::BlockNumber,
    },
    set_chain_availability_period {
        new: T::BlockNumber,
    },
    set_thread_availability_period {
        new: T::BlockNumber,
    },
    set_scheduling_lookahead {
        new: u32,
    },
    set_max_validators_per_core {
        new: Option<u32>,
    },
    set_max_validators {
        new: Option<u32>,
    },
    set_dispute_period {
        new: SessionIndex,
    },
    set_dispute_post_conclusion_acceptance_period {
        new: T::BlockNumber,
    },
    set_dispute_max_spam_slots {
        new: u32,
    },
    set_dispute_conclusion_by_time_out_period {
        new: T::BlockNumber,
    },
    set_no_show_slots {
        new: u32,
    },
    set_n_delay_tranches {
        new: u32,
    },
    set_zeroth_delay_tranche_width {
        new: u32,
    },
    set_needed_approvals {
        new: u32,
    },
    set_relay_vrf_modulo_samples {
        new: u32,
    },
    set_max_upward_queue_count {
        new: u32,
    },
    set_max_upward_queue_size {
        new: u32,
    },
    set_max_downward_message_size {
        new: u32,
    },
    set_ump_service_total_weight {
        new: Weight,
    },
    set_max_upward_message_size {
        new: u32,
    },
    set_max_upward_message_num_per_candidate {
        new: u32,
    },
    set_hrmp_open_request_ttl {
        new: u32,
    },
    set_hrmp_sender_deposit {
        new: Balance,
    },
    set_hrmp_recipient_deposit {
        new: Balance,
    },
    set_hrmp_channel_max_capacity {
        new: u32,
    },
    set_hrmp_channel_max_total_size {
        new: u32,
    },
    set_hrmp_max_parachain_inbound_channels {
        new: u32,
    },
    set_hrmp_max_parathread_inbound_channels {
        new: u32,
    },
    set_hrmp_channel_max_message_size {
        new: u32,
    },
    set_hrmp_max_parachain_outbound_channels {
        new: u32,
    },
    set_hrmp_max_parathread_outbound_channels {
        new: u32,
    },
    set_hrmp_max_message_num_per_candidate {
        new: u32,
    },
    set_ump_max_individual_weight {
        new: Weight,
    },
    // some variants omitted

}

Contains one variant per dispatchable that can be called by an extrinsic.

Variants

set_validation_upgrade_frequency

Set the validation upgrade frequency.

Fields of set_validation_upgrade_frequency

new: T::BlockNumber

set_validation_upgrade_delay

Set the validation upgrade delay.

Fields of set_validation_upgrade_delay

new: T::BlockNumber

set_code_retention_period

Set the acceptance period for an included candidate.

Fields of set_code_retention_period

new: T::BlockNumber

set_max_code_size

Set the max validation code size for incoming upgrades.

Fields of set_max_code_size

new: u32

set_max_pov_size

Set the max POV block size for incoming upgrades.

Fields of set_max_pov_size

new: u32

set_max_head_data_size

Set the max head data size for paras.

Fields of set_max_head_data_size

new: u32

set_parathread_cores

Set the number of parathread execution cores.

Fields of set_parathread_cores

new: u32

set_parathread_retries

Set the number of retries for a particular parathread.

Fields of set_parathread_retries

new: u32

set_group_rotation_frequency

Set the parachain validator-group rotation frequency

Fields of set_group_rotation_frequency

new: T::BlockNumber

set_chain_availability_period

Set the availability period for parachains.

Fields of set_chain_availability_period

new: T::BlockNumber

set_thread_availability_period

Set the availability period for parathreads.

Fields of set_thread_availability_period

new: T::BlockNumber

set_scheduling_lookahead

Set the scheduling lookahead, in expected number of blocks at peak throughput.

Fields of set_scheduling_lookahead

new: u32

set_max_validators_per_core

Set the maximum number of validators to assign to any core.

Fields of set_max_validators_per_core

new: Option<u32>

set_max_validators

Set the maximum number of validators to use in parachain consensus.

Fields of set_max_validators

new: Option<u32>

set_dispute_period

Set the dispute period, in number of sessions to keep for disputes.

Fields of set_dispute_period

new: SessionIndex

set_dispute_post_conclusion_acceptance_period

Set the dispute post conclusion acceptance period.

Fields of set_dispute_post_conclusion_acceptance_period

new: T::BlockNumber

set_dispute_max_spam_slots

Set the maximum number of dispute spam slots.

Fields of set_dispute_max_spam_slots

new: u32

set_dispute_conclusion_by_time_out_period

Set the dispute conclusion by time out period.

Fields of set_dispute_conclusion_by_time_out_period

new: T::BlockNumber

set_no_show_slots

Set the no show slots, in number of number of consensus slots. Must be at least 1.

Fields of set_no_show_slots

new: u32

set_n_delay_tranches

Set the total number of delay tranches.

Fields of set_n_delay_tranches

new: u32

set_zeroth_delay_tranche_width

Set the zeroth delay tranche width.

Fields of set_zeroth_delay_tranche_width

new: u32

set_needed_approvals

Set the number of validators needed to approve a block.

Fields of set_needed_approvals

new: u32

set_relay_vrf_modulo_samples

Set the number of samples to do of the RelayVRFModulo approval assignment criterion.

Fields of set_relay_vrf_modulo_samples

new: u32

set_max_upward_queue_count

Sets the maximum items that can present in a upward dispatch queue at once.

Fields of set_max_upward_queue_count

new: u32

set_max_upward_queue_size

Sets the maximum total size of items that can present in a upward dispatch queue at once.

Fields of set_max_upward_queue_size

new: u32

set_max_downward_message_size

Set the critical downward message size.

Fields of set_max_downward_message_size

new: u32

set_ump_service_total_weight

Sets the soft limit for the phase of dispatching dispatchable upward messages.

Fields of set_ump_service_total_weight

new: Weight

set_max_upward_message_size

Sets the maximum size of an upward message that can be sent by a candidate.

Fields of set_max_upward_message_size

new: u32

set_max_upward_message_num_per_candidate

Sets the maximum number of messages that a candidate can contain.

Fields of set_max_upward_message_num_per_candidate

new: u32

set_hrmp_open_request_ttl

Sets the number of sessions after which an HRMP open channel request expires.

Fields of set_hrmp_open_request_ttl

new: u32

set_hrmp_sender_deposit

Sets the amount of funds that the sender should provide for opening an HRMP channel.

Fields of set_hrmp_sender_deposit

new: Balance

set_hrmp_recipient_deposit

Sets the amount of funds that the recipient should provide for accepting opening an HRMP channel.

Fields of set_hrmp_recipient_deposit

new: Balance

set_hrmp_channel_max_capacity

Sets the maximum number of messages allowed in an HRMP channel at once.

Fields of set_hrmp_channel_max_capacity

new: u32

set_hrmp_channel_max_total_size

Sets the maximum total size of messages in bytes allowed in an HRMP channel at once.

Fields of set_hrmp_channel_max_total_size

new: u32

set_hrmp_max_parachain_inbound_channels

Sets the maximum number of inbound HRMP channels a parachain is allowed to accept.

Fields of set_hrmp_max_parachain_inbound_channels

new: u32

set_hrmp_max_parathread_inbound_channels

Sets the maximum number of inbound HRMP channels a parathread is allowed to accept.

Fields of set_hrmp_max_parathread_inbound_channels

new: u32

set_hrmp_channel_max_message_size

Sets the maximum size of a message that could ever be put into an HRMP channel.

Fields of set_hrmp_channel_max_message_size

new: u32

set_hrmp_max_parachain_outbound_channels

Sets the maximum number of outbound HRMP channels a parachain is allowed to open.

Fields of set_hrmp_max_parachain_outbound_channels

new: u32

set_hrmp_max_parathread_outbound_channels

Sets the maximum number of outbound HRMP channels a parathread is allowed to open.

Fields of set_hrmp_max_parathread_outbound_channels

new: u32

set_hrmp_max_message_num_per_candidate

Sets the maximum number of outbound HRMP messages can be sent by a candidate.

Fields of set_hrmp_max_message_num_per_candidate

new: u32

set_ump_max_individual_weight

Sets the maximum amount of weight any individual upward message may consume.

Fields of set_ump_max_individual_weight

new: Weight

Implementations

impl<T: Config> Call<T>

pub fn new_call_variant_set_validation_upgrade_frequency(
new: T::BlockNumber
) -> Self

Create a call with the variant set_validation_upgrade_frequency.

pub fn new_call_variant_set_validation_upgrade_delay(
new: T::BlockNumber
) -> Self

Create a call with the variant set_validation_upgrade_delay.

pub fn new_call_variant_set_code_retention_period(new: T::BlockNumber) -> Self

Create a call with the variant set_code_retention_period.

pub fn new_call_variant_set_max_code_size(new: u32) -> Self

Create a call with the variant set_max_code_size.

pub fn new_call_variant_set_max_pov_size(new: u32) -> Self

Create a call with the variant set_max_pov_size.

pub fn new_call_variant_set_max_head_data_size(new: u32) -> Self

Create a call with the variant set_max_head_data_size.

pub fn new_call_variant_set_parathread_cores(new: u32) -> Self

Create a call with the variant set_parathread_cores.

pub fn new_call_variant_set_parathread_retries(new: u32) -> Self

Create a call with the variant set_parathread_retries.

pub fn new_call_variant_set_group_rotation_frequency(
new: T::BlockNumber
) -> Self

Create a call with the variant set_group_rotation_frequency.

pub fn new_call_variant_set_chain_availability_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_chain_availability_period.

pub fn new_call_variant_set_thread_availability_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_thread_availability_period.

pub fn new_call_variant_set_scheduling_lookahead(new: u32) -> Self

Create a call with the variant set_scheduling_lookahead.

pub fn new_call_variant_set_max_validators_per_core(new: Option<u32>) -> Self

Create a call with the variant set_max_validators_per_core.

pub fn new_call_variant_set_max_validators(new: Option<u32>) -> Self

Create a call with the variant set_max_validators.

pub fn new_call_variant_set_dispute_period(new: SessionIndex) -> Self

Create a call with the variant set_dispute_period.

pub fn new_call_variant_set_dispute_post_conclusion_acceptance_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_dispute_post_conclusion_acceptance_period.

pub fn new_call_variant_set_dispute_max_spam_slots(new: u32) -> Self

Create a call with the variant set_dispute_max_spam_slots.

pub fn new_call_variant_set_dispute_conclusion_by_time_out_period(
new: T::BlockNumber
) -> Self

Create a call with the variant set_dispute_conclusion_by_time_out_period.

pub fn new_call_variant_set_no_show_slots(new: u32) -> Self

Create a call with the variant set_no_show_slots.

pub fn new_call_variant_set_n_delay_tranches(new: u32) -> Self

Create a call with the variant set_n_delay_tranches.

pub fn new_call_variant_set_zeroth_delay_tranche_width(new: u32) -> Self

Create a call with the variant set_zeroth_delay_tranche_width.

pub fn new_call_variant_set_needed_approvals(new: u32) -> Self

Create a call with the variant set_needed_approvals.

pub fn new_call_variant_set_relay_vrf_modulo_samples(new: u32) -> Self

Create a call with the variant set_relay_vrf_modulo_samples.

pub fn new_call_variant_set_max_upward_queue_count(new: u32) -> Self

Create a call with the variant set_max_upward_queue_count.

pub fn new_call_variant_set_max_upward_queue_size(new: u32) -> Self

Create a call with the variant set_max_upward_queue_size.

pub fn new_call_variant_set_max_downward_message_size(new: u32) -> Self

Create a call with the variant set_max_downward_message_size.

pub fn new_call_variant_set_ump_service_total_weight(new: Weight) -> Self

Create a call with the variant set_ump_service_total_weight.

pub fn new_call_variant_set_max_upward_message_size(new: u32) -> Self

Create a call with the variant set_max_upward_message_size.

pub fn new_call_variant_set_max_upward_message_num_per_candidate(
new: u32
) -> Self

Create a call with the variant set_max_upward_message_num_per_candidate.

pub fn new_call_variant_set_hrmp_open_request_ttl(new: u32) -> Self

Create a call with the variant set_hrmp_open_request_ttl.

pub fn new_call_variant_set_hrmp_sender_deposit(new: Balance) -> Self

Create a call with the variant set_hrmp_sender_deposit.

pub fn new_call_variant_set_hrmp_recipient_deposit(new: Balance) -> Self

Create a call with the variant set_hrmp_recipient_deposit.

pub fn new_call_variant_set_hrmp_channel_max_capacity(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_capacity.

pub fn new_call_variant_set_hrmp_channel_max_total_size(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_total_size.

pub fn new_call_variant_set_hrmp_max_parachain_inbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parachain_inbound_channels.

pub fn new_call_variant_set_hrmp_max_parathread_inbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parathread_inbound_channels.

pub fn new_call_variant_set_hrmp_channel_max_message_size(new: u32) -> Self

Create a call with the variant set_hrmp_channel_max_message_size.

pub fn new_call_variant_set_hrmp_max_parachain_outbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parachain_outbound_channels.

pub fn new_call_variant_set_hrmp_max_parathread_outbound_channels(
new: u32
) -> Self

Create a call with the variant set_hrmp_max_parathread_outbound_channels.

pub fn new_call_variant_set_hrmp_max_message_num_per_candidate(new: u32) -> Self

Create a call with the variant set_hrmp_max_message_num_per_candidate.

pub fn new_call_variant_set_ump_max_individual_weight(new: Weight) -> Self

Create a call with the variant set_ump_max_individual_weight.