Trait polkadot_runtime_parachains::disputes::slashing::pallet::Config

pub trait Config: Config + Config {
    type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount;
    type KeyOwnerIdentification: Parameter;
    type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, ValidatorId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>;
    type HandleReports: HandleReports<Self>;
    type WeightInfo: WeightInfo;
    type BenchmarkingConfig: BenchmarkingConfiguration;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type KeyOwnerProof: Parameter + GetSessionNumber + GetValidatorCount

The proof of key ownership, used for validating slashing reports. The proof must include the session index and validator count of the session at which the offence occurred.

type KeyOwnerIdentification: Parameter

The identification of a key owner, used when reporting slashes.

type KeyOwnerProofSystem: KeyOwnerProofSystem<(KeyTypeId, ValidatorId), Proof = Self::KeyOwnerProof, IdentificationTuple = Self::KeyOwnerIdentification>

A system for proving ownership of keys, i.e. that a given key was part of a validator set, needed for validating slashing reports.

type HandleReports: HandleReports<Self>

The slashing report handling subsystem, defines methods to report an offence (after the slashing report has been validated) and for submitting a transaction to report a slash (from an offchain context). NOTE: when enabling slashing report handling (i.e. this type isn’t set to ()) you must use this pallet’s ValidateUnsigned in the runtime definition.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type BenchmarkingConfig: BenchmarkingConfiguration

Benchmarking configuration.