Trait polkadot_runtime_parachains::disputes::PunishValidators

pub trait PunishValidators {
    fn punish_for_invalid(
        session: SessionIndex, 
        validators: impl IntoIterator<Item = ValidatorIndex>
    );

    fn punish_against_valid(
        session: SessionIndex, 
        validators: impl IntoIterator<Item = ValidatorIndex>
    );

    fn punish_inconclusive(
        session: SessionIndex, 
        validators: impl IntoIterator<Item = ValidatorIndex>
    );
}

Punishment hooks for disputes.

Required methods

fn punish_for_invalid(
session: SessionIndex,
validators: impl IntoIterator<Item = ValidatorIndex>
)

Punish a series of validators who were for an invalid parablock. This is expected to be a major punishment.

fn punish_against_valid(
session: SessionIndex,
validators: impl IntoIterator<Item = ValidatorIndex>
)

Punish a series of validators who were against a valid parablock. This is expected to be a minor punishment.

fn punish_inconclusive(
session: SessionIndex,
validators: impl IntoIterator<Item = ValidatorIndex>
)

Punish a series of validators who were part of a dispute which never concluded. This is expected to be a minor punishment.

Loading content...