Module polkadot_runtime_parachains::hrmp

Re-exports

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

HrmpChannel
A metadata of an HRMP channel.

HrmpOpenChannelRequest
A description of a request to open an HRMP channel.

Enums

HrmpWatermarkAcceptanceErr
An error returned by [check_hrmp_watermark] that indicates an acceptance criteria check didn’t pass.

OutboundHrmpAcceptanceErr
An error returned by [check_outbound_hrmp] that indicates an acceptance criteria check didn’t pass.