Module polkadot_runtime_parachains::hrmp::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_27 as tt_default_parts;

Structs

GenesisConfig

Preopen the given HRMP channels.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageHrmpAcceptedChannelRequestCount

_GeneratedPrefixForStorageHrmpChannelContents

_GeneratedPrefixForStorageHrmpChannelDigests

_GeneratedPrefixForStorageHrmpChannels

_GeneratedPrefixForStorageHrmpCloseChannelRequests

_GeneratedPrefixForStorageHrmpCloseChannelRequestsList

_GeneratedPrefixForStorageHrmpEgressChannelsIndex

_GeneratedPrefixForStorageHrmpIngressChannelsIndex

_GeneratedPrefixForStorageHrmpOpenChannelRequestCount

_GeneratedPrefixForStorageHrmpOpenChannelRequests

_GeneratedPrefixForStorageHrmpOpenChannelRequestsList

_GeneratedPrefixForStorageHrmpWatermarks

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

HrmpAcceptedChannelRequestCount

This mapping tracks how many open channel requests were accepted by a given recipient para. Invariant: HrmpOpenChannelRequests should contain the same number of items (_, X) with confirmed set to true, as the number of HrmpAcceptedChannelRequestCount for X.

HrmpChannelContents

Storage for the messages for each channel. Invariant: cannot be non-empty if the corresponding channel in HrmpChannels is None.

HrmpChannelDigests

Maintains a mapping that can be used to answer the question: What paras sent a message at the given block number for a given receiver. Invariants:

HrmpChannels

HRMP channel data associated with each para. Invariant:

HrmpCloseChannelRequests

A set of pending HRMP close channel requests that are going to be closed during the session change. Used for checking if a given channel is registered for closure.

HrmpCloseChannelRequestsList

HrmpEgressChannelsIndex

HrmpIngressChannelsIndex

Ingress/egress indexes allow to find all the senders and receivers given the opposite side. I.e.

HrmpOpenChannelRequestCount

This mapping tracks how many open channel requests are initiated by a given sender para. Invariant: HrmpOpenChannelRequests should contain the same number of items that has (X, _) as the number of HrmpOpenChannelRequestCount for X.

HrmpOpenChannelRequests

The set of pending HRMP open channel requests.

HrmpOpenChannelRequestsList

HrmpWatermarks

The HRMP watermark associated with each para. Invariant:

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.