Trait polkadot_runtime_parachains::hrmp::pallet::Config

pub trait Config: Config + Config + Config + Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Origin: From<Origin> + From<<Self as Config>::Origin> + Into<Result<Origin, <Self as Config>::Origin>>;
    type Currency: ReservableCurrency<Self::AccountId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

The outer event type.

type Origin: From<Origin> + From<<Self as Config>::Origin> + Into<Result<Origin, <Self as Config>::Origin>>

type Currency: ReservableCurrency<Self::AccountId>

An interface for reserving deposits for opening channels.

NOTE that this Currency instance will be charged with the amounts defined in the Configuration pallet. Specifically, that means that the Balance of the Currency implementation should be the same as Balance as used in the Configuration.