Module polkadot_runtime_parachains::inclusion

The inclusion pallet is responsible for inclusion and availability of scheduled parachains.

It is responsible for carrying candidates from being backable to being backed, and then from backed to included.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

AvailabilityBitfieldRecord

A bitfield signed by a validator indicating that it is keeping its piece of the erasure-coding for any backed candidates referred to by a 1 bit available.

CandidatePendingAvailability

A backed candidate pending availability.

TestWeightInfo

Enums

AggregateMessageOrigin

Aggregate message origin for the MessageQueue pallet.

UmpAcceptanceCheckErr

An error returned by [check_upward_messages] that indicates a violation of one of acceptance criteria rules.

UmpQueueId

Identifies a UMP queue inside the MessageQueue pallet.

Constants

MAX_UPWARD_MESSAGE_SIZE_BOUND

Maximum value that config.max_upward_message_size can be set to.

Traits

QueueFootprinter

Reads the footprint of queues for a specific origin type.

RewardValidators

A hook for applying validator rewards

WeightInfo

Functions

minimum_backing_votes

Number of backing votes we need for a valid backing.

Type Definitions

MaxUmpMessageLenOf

The maximal length of a UMP message.