Module polkadot_runtime_parachains::paras

The paras pallet is responsible for storing data on parachains and parathreads.

It tracks which paras are parachains, what their current head data is in this fork of the relay chain, what their validation code is, and what their past and upcoming validation code is.

A para is not considered live until it is registered and activated in this pallet. Activation can only occur at session boundaries.

Re-exports

pub use crate::Origin as ParachainOrigin;

pub use pallet::*;

Modules

pallet
The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

ParaGenesisArgs
Arguments for initializing a para.

ParaPastCodeMeta
Metadata used to track previous parachain validation code that we keep in the state.

ReplacementTimes

Enums

ParaLifecycle
The possible states of a para, to take into account delayed lifecycle changes.