Trait polkadot_runtime_parachains::paras::pallet::Config

pub trait Config: Config + Config + Config + CreateInherent<Call<Self>> {
    type RuntimeEvent: From<Event> + IsType<<Self as Config>::RuntimeEvent>;
    type UnsignedPriority: Get<TransactionPriority>;
    type NextSessionRotation: EstimateNextSessionRotation<BlockNumberFor<Self>>;
    type QueueFootprinter: QueueFootprinter<Origin = UmpQueueId>;
    type OnNewHead: OnNewHead;
    type WeightInfo: WeightInfo;
    type AssignCoretime: AssignCoretime;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event> + IsType<<Self as Config>::RuntimeEvent>

type UnsignedPriority: Get<TransactionPriority>

type NextSessionRotation: EstimateNextSessionRotation<BlockNumberFor<Self>>

type QueueFootprinter: QueueFootprinter<Origin = UmpQueueId>

Retrieve how many UMP messages are enqueued for this para-chain.

This is used to judge whether or not a para-chain can offboard. Per default this should be set to the ParaInclusion pallet.

type OnNewHead: OnNewHead

Runtime hook for when a parachain head is updated.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type AssignCoretime: AssignCoretime

Runtime hook for assigning coretime for a given parachain.

This is only used at genesis or by root.

TODO: Remove once coretime is the standard across all chains.

Object Safety

This trait is not object safe.