Trait polkadot_runtime_parachains::paras::AssignCoretime

pub trait AssignCoretime {
    // Required method
    fn assign_coretime(id: ParaId) -> DispatchResult;
}

Assign coretime to some parachain.

This assigns coretime to a parachain without using the coretime chain. Thus, this should only be used for testing purposes.

Required Methods

fn assign_coretime(id: ParaId) -> DispatchResult

ONLY USE FOR TESTING OR GENESIS.

Object Safety

This trait is not object safe.