Trait polkadot_runtime_parachains::paras_inherent::WeightInfo

pub trait WeightInfo {
    // Required methods
    fn enter_variable_disputes(v: u32) -> Weight;
    fn enter_bitfields() -> Weight;
    fn enter_backed_candidates_variable(v: u32) -> Weight;
    fn enter_backed_candidate_code_upgrade() -> Weight;
}

Required Methods

fn enter_variable_disputes(v: u32) -> Weight

Variant over v, the count of dispute statements in a dispute statement set. This gives the weight of a single dispute statement set.

fn enter_bitfields() -> Weight

The weight of one bitfield.

fn enter_backed_candidates_variable(v: u32) -> Weight

Variant over v, the count of validity votes for a backed candidate. This gives the weight of a single backed candidate.

fn enter_backed_candidate_code_upgrade() -> Weight

The weight of a single backed candidate with a code upgrade.

Object Safety

This trait is not object safe.