Module polkadot_runtime_parachains::runtime_api_impl::v1

A module exporting runtime API implementation functions for all runtime APIs using v1 primitives.

Runtimes implementing the v1 runtime API are recommended to forward directly to these functions.

Functions

availability_cores
Implementation for the availability_cores function of the runtime API.

candidate_events
Implementation for the candidate_events function of the runtime API.

candidate_pending_availability
Implementation for the candidate_pending_availability function of the runtime API.

check_validation_outputs
Implementation for the check_validation_outputs function of the runtime API.

dmq_contents
Implementation for the dmq_contents function of the runtime API.

inbound_hrmp_channels_contents
Implementation for the inbound_hrmp_channels_contents function of the runtime API.

persisted_validation_data
Implementation for the persisted_validation_data function of the runtime API.

relevant_authority_ids
Implementation for the AuthorityDiscoveryApi::authorities() function of the runtime API. It is a heavy call, but currently only used for authority discovery, so it is fine. Gets next, current and some historical authority ids using session_info module.

session_index_for_child
Implementation for the session_index_for_child function of the runtime API.

session_info
Get the session info for the given session, if stored.

validation_code
Implementation for the validation_code function of the runtime API.

validation_code_by_hash
Implementation for the validation_code_by_hash function of the runtime API.

validator_groups
Implementation for the validator_groups function of the runtime API.

validators
Implementation for the validators function of the runtime API.