Trait polkadot_runtime_parachains::scheduler::common::AssignmentProvider

pub trait AssignmentProvider<BlockNumber> {
    // Required methods
    fn pop_assignment_for_core(core_idx: CoreIndex) -> Option<Assignment>;
    fn report_processed(assignment: Assignment);
    fn push_back_assignment(assignment: Assignment);
    fn get_provider_config(
        core_idx: CoreIndex
    ) -> AssignmentProviderConfig<BlockNumber>;
    fn session_core_count() -> u32;
}

Required Methods

fn pop_assignment_for_core(core_idx: CoreIndex) -> Option<Assignment>

Pops an Assignment from the provider for a specified CoreIndex.

This is where assignments come into existance.

fn report_processed(assignment: Assignment)

A previously popped Assignment has been fully processed.

Report back to the assignment provider that an assignment is done and no longer present in the scheduler.

This is one way of the life of an assignment coming to an end.

fn push_back_assignment(assignment: Assignment)

Push back a previously popped assignment.

If the assignment could not be processed within the current session, it can be pushed back to the assignment provider in order to be poppped again later.

This is the second way the life of an assignment can come to an end.

fn get_provider_config( core_idx: CoreIndex ) -> AssignmentProviderConfig<BlockNumber>

Returns a set of variables needed by the scheduler

fn session_core_count() -> u32

How many cores are allocated to this provider.

As the name suggests the core count has to be session buffered:

Core count has to be predetermined for the next session in the current session.

Core count must not change during a session.